% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSunburstPlot.R
\name{addChild}
\alias{addChild}
\title{addChild}
\usage{
addChild(j, children, parts, root)
}
\arguments{
\item{j}{iterator}

\item{children}{children to add}

\item{parts}{labels of treatments}

\item{root}{root list}
}
\value{
root list with added childs
}
\description{
addChild
}
