% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructPathways.R
\name{addLabels}
\alias{addLabels}
\title{addLabels}
\usage{
addLabels(treatmentHistory, outputFolder)
}
\arguments{
\item{treatmentHistory}{treatmentHistory object}

\item{outputFolder}{Folder of output}
}
\value{
treatmentHistory
}
\description{
Adds back cohort names to concept ids.
}
