% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSettings.R
\name{addPathwaySettings}
\alias{addPathwaySettings}
\title{Add set of pathway settings.}
\usage{
addPathwaySettings(
  studyName = c("default"),
  targetCohortId,
  eventCohortIds,
  includeTreatmentsPriorToIndex = 0,
  minEraDuration = 0,
  splitEventCohorts = "",
  eraCollapseSize = 0,
  combinationWindow = 30,
  minStepDuration = 30,
  filterTreatments = "First",
  maxPathLength = 5,
  minCellCount = 0,
  minCellMethod = "Remove",
  groupCombinations = 10,
  addNoPaths = FALSE
)
}
\arguments{
\item{studyName}{Name identifying the set of study parameters.}

\item{targetCohortId}{Target cohort ID of current study settings.}

\item{eventCohortIds}{Event cohort IDs of current study settings.}

\item{includeTreatmentsPriorToIndex}{# Number of days prior to the index date of the target cohort that event cohorts are allowed to start}

\item{minEraDuration}{# Minimum time an event era should last to be included in analysis}

\item{splitEventCohorts}{# Specify event cohort to split in acute (< 30 days) and therapy (>= 30 days)}

\item{eraCollapseSize}{# Window of time between which two eras of the same event cohort are collapsed into one era}

\item{combinationWindow}{# Window of time two event cohorts need to overlap to be considered a combination treatment}

\item{minStepDuration}{# Minimum time an event era before or after a generated combination treatment should last to be included in analysis}

\item{filterTreatments}{# Select first occurrence of / changes between / all event cohorts}

\item{maxPathLength}{# Maximum number of steps included in treatment pathway (max 5)}

\item{minCellCount}{# Minimum number of persons with a specific treatment pathway for the pathway to be included in analysis}

\item{minCellMethod}{# Select to completely remove / sequentially adjust (by removing last step as often as necessary) treatment pathways below minCellCount}

\item{groupCombinations}{# Select to group all non-fixed combinations in one category 'other’ in the sunburst plot}

\item{addNoPaths}{# Select to include untreated persons without treatment pathway in the sunburst plot}
}
\value{

}
\description{
Add set of pathway settings.
}
