% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPathwaySettings.R
\name{addPathwaySettingsCheck}
\alias{addPathwaySettingsCheck}
\title{addPathwaySettingsCheck}
\usage{
addPathwaySettingsCheck(
  studyName,
  targetCohortId,
  eventCohortIds,
  includeTreatments,
  periodPriorToIndex,
  minEraDuration,
  splitEventCohorts,
  splitTime,
  eraCollapseSize,
  combinationWindow,
  minPostCombinationDuration,
  filterTreatments,
  maxPathLength,
  minCellCount,
  minCellMethod,
  groupCombinations,
  addNoPaths
)
}
\arguments{
\item{addNoPaths}{}
}
\value{
TRUE if all assertions pass
}
\description{
Helper functions that checks the input for addPathwaySettings.
}
