% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-AnalysisSettings.R
\name{AnalysisSettings}
\alias{AnalysisSettings}
\title{AnalysisSettings}
\description{
R6 AnalysisSettings object.
}
\examples{
cohortSettings <- CohortSettings$new(
  targetCohorts = data.frame(cohortId = c(1), cohortName = c("disease")),
  eventCohorts = data.frame(cohortId = c(2, 3), cohortName = c("drugA", "drugB"))
)

AnalysisA <- AnalysisSettings$new(
  cohortSettings = cohortSettings,
  studyName = "Analysis 1"
)
}
\seealso{
Other Settings: 
\code{\link{CohortSettings}},
\code{\link{DataSettings}},
\code{\link{PathwaySettings}},
\code{\link{SaveSettings}}
}
\concept{Settings}
\section{Super class}{
\code{TreatmentPatterns::Settings} -> \code{AnalysisSettings}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnalysisSettings-new}{\code{AnalysisSettings$new()}}
\item \href{#method-AnalysisSettings-editSettings}{\code{AnalysisSettings$editSettings()}}
\item \href{#method-AnalysisSettings-get}{\code{AnalysisSettings$get()}}
\item \href{#method-AnalysisSettings-validate}{\code{AnalysisSettings$validate()}}
\item \href{#method-AnalysisSettings-clone}{\code{AnalysisSettings$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="TreatmentPatterns" data-topic="Settings" data-id="print"><a href='../../TreatmentPatterns/html/Settings.html#method-Settings-print'><code>TreatmentPatterns::Settings$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnalysisSettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnalysisSettings-new}{}}}
\subsection{Method \code{new()}}{
Initializer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalysisSettings$new(cohortSettings, studyName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cohortSettings}}{(\code{CohortSettings})
An instance of CohortSettings R6 object.}

\item{\code{studyName}}{(\link[base]{character})
Name of the study.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnalysisSettings-editSettings"></a>}}
\if{latex}{\out{\hypertarget{method-AnalysisSettings-editSettings}{}}}
\subsection{Method \code{editSettings()}}{
Method to edit the analysis settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalysisSettings$editSettings(
  includeTreatments = "startDate",
  periodPriorToIndex = 0,
  minEraDuration = 0,
  splitEventCohorts = "",
  splitTime = 30,
  eraCollapseSize = 30,
  combinationWindow = 30,
  minPostCombinationDuration = 30,
  filterTreatments = "First",
  maxPathLength = 5,
  minCellCount = 5,
  minCellMethod = "Remove",
  groupCombinations = 10,
  addNoPaths = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{includeTreatments}}{(\link[base]{character})}

\item{\code{periodPriorToIndex}}{(\link[base]{numeric})}

\item{\code{minEraDuration}}{(\link[base]{numeric})}

\item{\code{splitEventCohorts}}{(\link[base]{character})}

\item{\code{splitTime}}{(\link[base]{numeric})}

\item{\code{eraCollapseSize}}{(\link[base]{numeric})}

\item{\code{combinationWindow}}{(\link[base]{numeric})}

\item{\code{minPostCombinationDuration}}{(\link[base]{numeric})}

\item{\code{filterTreatments}}{(\link[base]{character})}

\item{\code{maxPathLength}}{(\link[base]{numeric})}

\item{\code{minCellCount}}{(\link[base]{numeric})}

\item{\code{minCellMethod}}{(\link[base]{character})}

\item{\code{groupCombinations}}{(\link[base]{numeric})}

\item{\code{addNoPaths}}{(\link[base]{logical})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnalysisSettings-get"></a>}}
\if{latex}{\out{\hypertarget{method-AnalysisSettings-get}{}}}
\subsection{Method \code{get()}}{
Gettter method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalysisSettings$get()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\link[base]{list})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnalysisSettings-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AnalysisSettings-validate}{}}}
\subsection{Method \code{validate()}}{
validate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalysisSettings$validate(settings)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{settings}}{(\link[base]{list})
List of settings}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnalysisSettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnalysisSettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalysisSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
