% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{appendCovariateData}
\alias{appendCovariateData}
\title{appendCovariateData}
\usage{
appendCovariateData(tempCovariateData, covariateData)
}
\arguments{
\item{tempCovariateData}{Temporary covariate data}

\item{covariateData}{Covariate data}
}
\value{
covariateData
    Covariate data
}
\description{
Borrowed function from Triton.
}
