% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSunburstPlot.R
\name{buildHierarchy}
\alias{buildHierarchy}
\title{buildHierarchy}
\usage{
buildHierarchy(csv)
}
\arguments{
\item{csv}{CSV}
}
\value{
JSON
}
\description{
Help function to create hierarchical data structure.
}
