% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-CDMSettings.R
\name{CDMSettings}
\alias{CDMSettings}
\title{R6 CDMSettings class}
\description{
CDMSettings object. Inherits from Settings
}
\examples{
# Using a ConnectionDetails
connectionDetails <- Eunomia::getConnectionDetails()

CDMSettings$new(
  connectionDetails = connectionDetails,
  cdmDatabaseSchema = "main",
  resultSchema = "main",
  cohortTable = "cohortTable"
)

# Using a CDMConnector cdm object.
con <- DBI::dbConnect(duckdb::duckdb(), dbdir = eunomia_dir())
cdm <- CDMConnector::cdmFromCon(con, cdmSchema = "main")

CDMSettings$new(
  cdm = cdm,
  cohortTable = "cohortTable"
)
}
\concept{Settings}
\section{Super class}{
\code{\link[TreatmentPatterns:Settings]{TreatmentPatterns::Settings}} -> \code{CDMSettings}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CDMSettings-new}{\code{CDMSettings$new()}}
\item \href{#method-CDMSettings-print}{\code{CDMSettings$print()}}
\item \href{#method-CDMSettings-validate}{\code{CDMSettings$validate()}}
\item \href{#method-CDMSettings-get}{\code{CDMSettings$get()}}
\item \href{#method-CDMSettings-fetchCohortTable}{\code{CDMSettings$fetchCohortTable()}}
\item \href{#method-CDMSettings-clone}{\code{CDMSettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CDMSettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-CDMSettings-new}{}}}
\subsection{Method \code{new()}}{
!Overload initialize new
Initializer method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDMSettings$new(
  cohortTable,
  connectionDetails = NULL,
  cdmDatabaseSchema = "",
  resultSchema = "",
  cdm = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CDMSettings-print"></a>}}
\if{latex}{\out{\hypertarget{method-CDMSettings-print}{}}}
\subsection{Method \code{print()}}{
!Overload print
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDMSettings$print(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{CDMSettings object.}

\item{\code{...}}{Other parameters for \link[base]{print}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Message from \link[base]{cat}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CDMSettings-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CDMSettings-validate}{}}}
\subsection{Method \code{validate()}}{
!overload validation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDMSettings$validate()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connectionDetails}}{\link[DatabaseConnector]{createConnectionDetails}\cr
ConnectionDetails object.}

\item{\code{cdmDatabaseSchema}}{\link[base]{character}\cr
CDM Schema in database.}

\item{\code{resultSchema}}{\link[base]{character}\cr
Result Schema in database.}

\item{\code{cohortTable}}{\link[base]{character}\cr
Name of the cohort table.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CDMSettings-get"></a>}}
\if{latex}{\out{\hypertarget{method-CDMSettings-get}{}}}
\subsection{Method \code{get()}}{
Gets the specified items packged in a named list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDMSettings$get()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\link[base]{list}
\describe{
\item{connectionDetails}{(\code{ConnectionDetails})ConnectionDetails object. See \link[DatabaseConnector]{createConnectionDetails}.}
\item{cdmDatabaseSchema}{(\link[base]{character}) Name of the CDM Schema in database.}
\item{resultSchema}{(\link[base]{character}) Name of the Result Schema in the database.}
\item{cohortTable}{(\link[base]{character}) Name of the cohort table}
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CDMSettings-fetchCohortTable"></a>}}
\if{latex}{\out{\hypertarget{method-CDMSettings-fetchCohortTable}{}}}
\subsection{Method \code{fetchCohortTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDMSettings$fetchCohortTable(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CDMSettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CDMSettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDMSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
