% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPathwaySettings.R
\name{checkAddPathwaySettings}
\alias{checkAddPathwaySettings}
\title{checkAddPathwaySettings}
\usage{
checkAddPathwaySettings(env)
}
\arguments{
\item{env}{Environment containging all the function environment variables.}
}
\value{
TRUE if all assertions pass
}
\description{
Asserts that parameters fall within specified bounds.
}
\examples{
\dontrun{
TreatmentPatterns:::checkAddPathwaySettings(env)
}
}
