% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructPathways.R
\name{checkConstructPathways}
\alias{checkConstructPathways}
\title{checkConstructPathways}
\usage{
checkConstructPathways(env)
}
\arguments{
\item{env}{Environment containging all the function environment variables.}
}
\value{
TRUE if all assertions pass
}
\description{
Checks parameters for constructPathways.
}
