% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDataSettings.R
\name{checkDataSettings}
\alias{checkDataSettings}
\title{checkDataSettings}
\usage{
checkDataSettings(env)
}
\arguments{
\item{env}{Environment containging all the function environment variables.}
}
\value{
TRUE if all the assertions pass.
}
\description{
Check createDataSettings parameters.
}
