% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPathwaySettings.R
\name{checkPathwaySettings}
\alias{checkPathwaySettings}
\title{checkPathwaySettings}
\usage{
checkPathwaySettings(cohortSettings)
}
\arguments{
\item{cohortSettings}{cohortSettings object}
}
\value{
TRUE if all assertions pass
}
\description{
Check createPathwaySettings
}
