% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSaveSettings.R
\name{checkSaveSettings}
\alias{checkSaveSettings}
\title{checkSaveSettings}
\usage{
checkSaveSettings(databaseName, rootFolder, outputFolder, tempFolder)
}
\arguments{
\item{databaseName}{}

\item{rootFolder}{}

\item{outputFolder}{}

\item{tempFolder}{}
}
\value{
TRUE if all assertions pass
}
\description{
Checks parameters of createSaveSettings.
}
