% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortSettings.R
\name{CohortSettings}
\alias{CohortSettings}
\title{CohortSettings}
\description{
CohortSettings

CohortSettings
}
\details{
CohortSettings object, inherits from Settings.
}
\section{Super class}{
\code{TreatmentPatterns::Settings} -> \code{CohortSettings}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CohortSettings-new}{\code{CohortSettings$new()}}
\item \href{#method-CohortSettings-get}{\code{CohortSettings$get()}}
\item \href{#method-CohortSettings-getTarget}{\code{CohortSettings$getTarget()}}
\item \href{#method-CohortSettings-getEvent}{\code{CohortSettings$getEvent()}}
\item \href{#method-CohortSettings-getExit}{\code{CohortSettings$getExit()}}
\item \href{#method-CohortSettings-validate}{\code{CohortSettings$validate()}}
\item \href{#method-CohortSettings-clone}{\code{CohortSettings$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="TreatmentPatterns" data-topic="Settings" data-id="print"><a href='../../TreatmentPatterns/html/Settings.html#method-Settings-print'><code>TreatmentPatterns::Settings$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortSettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-CohortSettings-new}{}}}
\subsection{Method \code{new()}}{
initialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSettings$new(
  eventCohorts = NULL,
  targetCohorts = NULL,
  exitCohorts = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{eventCohorts}}{<data.frame> Data frame containing event cohorts. Should have two
columns: 1) cohortId, 2) cohortName.}

\item{\code{targetCohorts}}{<data.frame> Data frame containing target cohorts. Should have two
columns: 1) cohortId, 2) cohortName.}

\item{\code{exitCohorts}}{<data.frame> Data frame containing exit cohorts. Should have two
columns: 1) cohortId, 2) cohortName.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(self)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortSettings-get"></a>}}
\if{latex}{\out{\hypertarget{method-CohortSettings-get}{}}}
\subsection{Method \code{get()}}{
get
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSettings$get()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<data.frame>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortSettings-getTarget"></a>}}
\if{latex}{\out{\hypertarget{method-CohortSettings-getTarget}{}}}
\subsection{Method \code{getTarget()}}{
getTarget
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSettings$getTarget()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<data.frame>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortSettings-getEvent"></a>}}
\if{latex}{\out{\hypertarget{method-CohortSettings-getEvent}{}}}
\subsection{Method \code{getEvent()}}{
getEvent
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSettings$getEvent()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<data.frame>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortSettings-getExit"></a>}}
\if{latex}{\out{\hypertarget{method-CohortSettings-getExit}{}}}
\subsection{Method \code{getExit()}}{
getExit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSettings$getExit()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<data.frame>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortSettings-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CohortSettings-validate}{}}}
\subsection{Method \code{validate()}}{
validate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSettings$validate(cohortsList)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cohortsList}}{<list> List of target, event, and exit cohort data frames.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(self)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortSettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CohortSettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
