% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructPathways.R
\name{combineData}
\alias{combineData}
\title{combineData}
\usage{
combineData(andromeda, iterations)
}
\arguments{
\item{andromeda}{(\code{Andromeda::andromeda()})}

\item{iterations}{(\code{integer(1)})}
}
\value{
(\code{invisible(NULL)})
}
\description{
combineData
}
