% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{computeCounts}
\alias{computeCounts}
\title{computeCounts}
\usage{
computeCounts(treatmentHistory, minFreq)
}
\arguments{
\item{treatmentHistory}{(\code{data.frame()})}
}
\value{
(\code{list()})
}
\description{
computeCounts
}
