% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{computeTreatmentPathways}
\alias{computeTreatmentPathways}
\title{computeTreatmentPathways}
\usage{
computeTreatmentPathways(treatmentHistory, ageWindow, minCellCount, censorType)
}
\arguments{
\item{treatmentHistory}{data.frame()}

\item{ageWindow}{numeric(n)}

\item{minCellCount}{numeric(1)}

\item{censorType}{character(1)}
}
\value{
(\code{data.frame()})
}
\description{
computeTreatmentPathways
}
