% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructPathways.R
\name{constructPathway}
\alias{constructPathway}
\title{constructPathway}
\usage{
constructPathway(settings, cohortParams, andromeda, localAndromeda)
}
\arguments{
\item{settings}{(\code{data.frame})}

\item{cohortParams}{(\code{data.frame})}

\item{andromeda}{(\code{Andromeda::andromeda()})}

\item{localAndromeda}{(\code{Andromeda::andromeda()})}
}
\value{
\code{invisible(NULL)}
}
\description{
Constructs the pathways.
}
