% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSettings.R
\name{createCharacterizationSettings}
\alias{createCharacterizationSettings}
\title{Create characterization settings (optional, only for OMOP-CDM data ).}
\usage{
createCharacterizationSettings(
  baselineCovariates_location = NULL,
  baselineCovariates = data.frame(covariateName = c("Male", "Age",
    "Charlson comorbidity index score"), covariateId = c(8507001, 1002, 1901)),
  standardCovariateSettings =
    FeatureExtraction::createCovariateSettings(useDemographicsAge = TRUE,
    useDemographicsGender = TRUE, useDemographicsTimeInCohort = TRUE,
    useDemographicsPostObservationTime = TRUE, useConditionGroupEraAnyTimePrior = TRUE,
    useConditionGroupEraLongTerm = TRUE, useCharlsonIndex = TRUE),
  returnCovariates = "all",
  minCellCount = 5
)
}
\arguments{
\item{baselineCovariates_location}{Optional: Location of saved baselineCovariates object.}

\item{baselineCovariates}{Data frame containing the baseline characteristics of interest
(covariateName = "Descriptive name covariate", covariateId = "Unique ID
number referring to covariate from FeatureExtraction or 'Custom' (see
explanation below)"), covariateId can be "custom" if SQL code is.}

\item{standardCovariateSettings}{An object of type covariateSettings as created using the
createCovariateSettings function in the FeatureExtraction package.}

\item{returnCovariates}{Return "all" features or only "selection" of features}

\item{minCellCount}{Minimum number of persons with a specific baseline covariate to be
included in analysis}
}
\value{
Object characterizationSettings.
}
\description{
Create characterization settings (optional, only for OMOP-CDM data ).
}
