% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCohort.R
\name{createCohorts}
\alias{createCohorts}
\title{This function will create the target and event cohorts following the definitions included in
this package if OMOP-CDM = TRUE.}
\usage{
createCohorts(
  connection,
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  outputFolder,
  instFolder,
  loadCohorts = FALSE,
  baseUrl,
  generateCohorts = TRUE,
  minCellCount,
  flowChart = TRUE
)
}
\arguments{
\item{connection}{Connection to database server.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data resides if OMOP-CDM = TRUE.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{cohortTable}{The name of the table that will be created in the work database schema.
This table will hold the target and event cohorts used in this
study.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/).}

\item{loadCohorts}{Setting to load cohorts from ATLAS.}

\item{baseUrl}{The base URL for theWebApi instance, for example: "http://server.org:80/WebAPI".
Note, there is no trailing '/'. If trailing '/' is used, you may receive an error.}

\item{generateCohorts}{Setting to extract specified target/event cohorts from database.}

\item{minCellCount}{Minimum number of persons with a specific treatment pathway for the pathway to be included in analysis.}

\item{flowChart}{Setting to return numbers for flowchart with inclusion/exclusion criteria.}
}
\description{
This function will create the target and event cohorts following the definitions included in
this package if OMOP-CDM = TRUE.
}
