% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{createCustomCovariateSettings}
\alias{createCustomCovariateSettings}
\title{createCustomCovariateSettings}
\usage{
createCustomCovariateSettings(list_covariates)
}
\arguments{
\item{list_covariates}{List of covariates}
}
\value{
covariateSettings
}
\description{
createCustomCovariateSettings
}
