% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOutput.R
\name{createLegend}
\alias{createLegend}
\title{createLegend}
\usage{
createLegend(studyName, outputFolder, databaseName)
}
\arguments{
\item{studyName}{Study name}

\item{outputFolder}{Output folder}

\item{databaseName}{Database name}
}
\description{
createLegend
}
