% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPathwaySettings.R
\name{createPathwaySettings}
\alias{createPathwaySettings}
\title{createPathwaySettings}
\usage{
createPathwaySettings(cohortSettings, ...)
}
\arguments{
\item{cohortSettings}{cohortSettings object}

\item{...}{Any addPathwaySettings parameter:
1. studyName
2. targetCohortId
3. eventCohortIds
4. includeTreatments
5. periodPriorToIndex
6. minEraDuration
7. splitEventCohorts
8. splitTime
9. eraCollapseSize
10. combinationWindow
11. minPostCombinationDuration
12. filterTreatments
13. maxPathLength
14. minCellCount
15. minCellMethod
16. groupCombinations
17. addNoPaths}
}
\value{
Object pathwaySettings.
}
\description{
Create pathway settings.
}
\examples{
createPathwaySettings(
  cohortSettings = cohortSettings,
  studyName = "MyStudyName")
}
