% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSettings.R
\name{createSaveSettings}
\alias{createSaveSettings}
\title{Create save settings.}
\usage{
createSaveSettings(
  databaseName = "unknown_name",
  rootFolder = getwd(),
  outputFolder = paste0(rootFolder, "/output"),
  tempFolder = paste0(rootFolder, "/temp")
)
}
\arguments{
\item{databaseName}{Name of the database that will appear in the results.}

\item{rootFolder}{Name of local folder to place all package output (outputFolder, tempFolder if not given).}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes (/).}

\item{tempFolder}{Name of local folder to place intermediate results (not to be shared); make sure to use forward slashes (/).}
}
\description{
Create save settings.
}
