% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStudySettings.R
\name{createStudySettings}
\alias{createStudySettings}
\title{Transforms and saves input arguments given by user in instFolder/settings.}
\usage{
createStudySettings(
  instFolder,
  targetCohorts = NULL,
  eventCohorts = NULL,
  characterizationSettings = NULL,
  pathwaySettings = NULL
)
}
\arguments{
\item{instFolder}{Name of local folder to place all settings and cohorts; make sure to use forward slashes (/).}

\item{targetCohorts}{Data frame containing the study population of interest (cohortId, atlasId, cohortName).}

\item{eventCohorts}{Data frame containing the events of interest (cohortId, atlasId, cohortName).}

\item{characterizationSettings}{Data frame containing the baseline characteristics of interest (covariateName, covariateId).}

\item{pathwaySettings}{Data frame containing all pathway settings.}
}
\description{
Transforms and saves input arguments given by user in instFolder/settings.
}
