% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSunburstPlot.R
\name{createTreatmentPathways}
\alias{createTreatmentPathways}
\title{createTreatmentPathways}
\usage{
createTreatmentPathways(treatmentHistory)
}
\arguments{
\item{treatmentHistory}{(\code{data.frame()})}
}
\value{
(\code{data.frame()})
}
\description{
createTreatmentPathways
}
