% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darwinExecute.R
\name{darwinExecute}
\alias{darwinExecute}
\title{darwinExecute}
\usage{
darwinExecute(
  targetCohort,
  cohortPath,
  dataSettings,
  pathwaySettings,
  characterizatonSettings,
  saveSettings
)
}
\arguments{
\item{targetCohort}{}

\item{cohortPath}{}

\item{dataSettings}{}

\item{pathwaySettings}{}

\item{characterizatonSettings}{}

\item{saveSettings}{}
}
\description{
darwinExecute
}
