% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSettings.R
\name{DataSettings}
\alias{DataSettings}
\title{DataSettings}
\description{
DataSettings

DataSettings
}
\details{
DataSettings R6 object. Inherits from Settings
}
\section{Super class}{
\code{TreatmentPatterns::Settings} -> \code{DataSettings}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataSettings-new}{\code{DataSettings$new()}}
\item \href{#method-DataSettings-print}{\code{DataSettings$print()}}
\item \href{#method-DataSettings-get}{\code{DataSettings$get()}}
\item \href{#method-DataSettings-validate}{\code{DataSettings$validate()}}
\item \href{#method-DataSettings-clone}{\code{DataSettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataSettings-new}{}}}
\subsection{Method \code{new()}}{
initialize method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSettings$new(
  connectionDetails,
  cdmDatabaseSchema = "",
  resultSchema = "",
  cohortTable = ""
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connectionDetails}}{<ConnectionDetails> object >= 6.0.0.}

\item{\code{cdmDatabaseSchema}}{<character> CDM schema name in the database.}

\item{\code{resultSchema}}{<character> Name of results schema.}

\item{\code{cohortTable}}{<character> Name of the cohort table in the results schema.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(self)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSettings-print"></a>}}
\if{latex}{\out{\hypertarget{method-DataSettings-print}{}}}
\subsection{Method \code{print()}}{
print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSettings$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible(self)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSettings-get"></a>}}
\if{latex}{\out{\hypertarget{method-DataSettings-get}{}}}
\subsection{Method \code{get()}}{
get method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSettings$get()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<list>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSettings-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DataSettings-validate}{}}}
\subsection{Method \code{validate()}}{
validate method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSettings$validate(
  connectionDetails,
  cdmDatabaseSchema,
  resultSchema,
  cohortTable
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connectionDetails}}{<ConnectionDetails> object >= 6.0.0.}

\item{\code{cdmDatabaseSchema}}{<character> CDM schema name in the database.}

\item{\code{resultSchema}}{<character> Name of results schema.}

\item{\code{cohortTable}}{<character> Name of the cohort table in the results schema.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(self)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataSettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
