% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSunburstPlot.R
\name{depth}
\alias{depth}
\title{depth}
\usage{
depth(x, thisdepth = 0)
}
\arguments{
\item{x}{input list (element)}

\item{thisdepth}{current list depth}
}
\value{
the depth of the list element
}
\description{
Function to find depth of a list element.
}
