% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructPathways.R
\name{doCombinationWindow}
\alias{doCombinationWindow}
\title{doCombinationWindow}
\usage{
doCombinationWindow(
  treatment_history,
  combinationWindow,
  minPostCombinationDuration
)
}
\arguments{
\item{treatment_history}{Dataframe with event cohorts of the target cohort in different rows.}

\item{combinationWindow}{Window of time two event cohorts need to overlap to be considered a
combination treatment.}

\item{minPostCombinationDuration}{Minimum time an event era before or after a generated combination
treatment should last to be included in analysis.}
}
\value{
treatment_history
    Updated dataframe, where overlapping event cohorts 
    are modified according to rules defined for switching / combinations.
}
\description{
doCombinationWindow
}
