% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructPathways.R
\name{doCreateTreatmentHistory}
\alias{doCreateTreatmentHistory}
\title{doCreateTreatmentHistory}
\usage{
doCreateTreatmentHistory(
  currentCohorts,
  targetCohortId,
  eventCohortIds,
  periodPriorToIndex,
  includeTreatments
)
}
\arguments{
\item{currentCohorts}{Dataframe with target and event cohorts of current study settings.}

\item{targetCohortId}{Target cohort ID of current study settings.}

\item{eventCohortIds}{Event cohort IDs of current study settings.}

\item{periodPriorToIndex}{Number of days prior to the index date of the target cohort that event
cohorts are allowed to start}

\item{includeTreatments}{Include treatments starting ('startDate') or ending ('endDate') after
target cohort start date}
}
\value{
currentCohorts
    Updated dataframe, including only event cohorts after
    target cohort start date and with added index year, duration, gap same
    columns.
}
\description{
doCreateTreatmentHistory
}
