% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructPathways.R
\name{doEraCollapseNew}
\alias{doEraCollapseNew}
\title{doEraCollapseNew}
\usage{
doEraCollapseNew(andromeda, eraCollapseSize)
}
\arguments{
\item{andromeda}{(\code{Andromeda::andromeda()})}

\item{eraCollapseSize}{(\code{integer(1)})}
}
\value{
\code{NULL}
}
\description{
doEraCollapseNew
}
