% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructPathways.R
\name{doEraDuration}
\alias{doEraDuration}
\title{doEraDuration}
\usage{
doEraDuration(treatment_history, minEraDuration)
}
\arguments{
\item{treatment_history}{Dataframe with event cohorts of the target cohort in different rows.}

\item{minEraDuration}{Minimum time an event era should last to be included in analysis.}
}
\value{
treatment_history
    Updated dataframe, rows with duration < 
    minEraDuration filtered out.
}
\description{
doEraDuration
}
