% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructPathways.R
\name{doMaxPathLength}
\alias{doMaxPathLength}
\title{doMaxPathLength}
\usage{
doMaxPathLength(treatmentHistory, maxPathLength)
}
\arguments{
\item{treatmentHistory}{Dataframe with event cohorts of the target cohort in different rows.}

\item{maxPathLength}{Maximum number of steps included in treatment pathway.}
}
\value{
treatmentHistory
Updated dataframe, where the desired event cohorts all have a seq value of <=
maxPathLength
}
\description{
Filters the treatmentHistory data.frame where eventSeq is smaller or equal
than maxPathLength
}
\examples{
\dontrun{
th <- doCreateTreatmentHistory(current_cohorts = currentCohorts,
                               targetCohortId = targetCohortId,
                               eventCohortIds = eventCohortIds,
                               periodPriorToIndex = periodPriorToIndex,
                               includeTreatments = includeTreatments)
doMaxPathLength(treatmentHistory = th, maxPathLength = 1)}
}
