% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOutput.R
\name{doMinCellCount}
\alias{doMinCellCount}
\title{doMinCellCount}
\usage{
doMinCellCount(
  file_noyear,
  file_withyear,
  outputFolder,
  tempFolder,
  databaseName,
  studyName,
  groupCombinations,
  minCellCount,
  minCellMethod
)
}
\arguments{
\item{file_noyear}{Dataframe with aggregated treatment pathways of the target cohort in
different rows (unique pathways, with frequency).}

\item{file_withyear}{Idem to file_noyear, but split over index_year.}

\item{outputFolder}{Path to the output folder.}

\item{tempFolder}{Path to the temp folder.}

\item{databaseName}{Name of the database that will appear in the results.}

\item{studyName}{Name for the study corresponding to the current settings.}

\item{groupCombinations}{Select to group all non-fixed combinations in one category "other" in
the sunburst plot.}

\item{minCellCount}{Minimum number of persons with a specific treatment pathway for the
pathway to be included in analysis.}

\item{minCellMethod}{Select to completely remove / sequentially adjust (by removing last step
as often as necessary) treatment pathways below minCellCount.}
}
\value{
List
    List of file_noyear and file_withyear

List
}
\description{
Apply minCellCount to generate censored treatment pathways to share and for
generating detailed output.
}
