% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructPathways.R
\name{doStepDuration}
\alias{doStepDuration}
\title{doStepDuration}
\usage{
doStepDuration(treatment_history, minPostCombinationDuration)
}
\arguments{
\item{treatment_history}{Dataframe with event cohorts of the target cohort in different rows.}

\item{minPostCombinationDuration}{Minimum time an event era before or after a generated combination 
treatment should last to be included in analysis.}
}
\value{
treatment_history
    Updated dataframe, rows with duration_era < 
    minPostCombinationDuration filtered out.
}
\description{
doStepDuration
}
