% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{exportCharacterization}
\alias{exportCharacterization}
\title{exportCharacterization}
\usage{
exportCharacterization(
  characteristics,
  databaseId,
  incremental,
  covariateValueFileName,
  covariateRefFileName,
  analysisRefFileName,
  timeRefFileName = NULL,
  counts,
  minCellCount
)
}
\arguments{
\item{characteristics}{Characteristics}

\item{databaseId}{Database ID}

\item{incremental}{Incremental}

\item{covariateValueFileName}{Covariate value file name}

\item{covariateRefFileName}{Covariate reference file name}

\item{analysisRefFileName}{Analysis reference file name}

\item{timeRefFileName}{Time reference file name}

\item{counts}{Counts}

\item{minCellCount}{Minimum cell count}
}
\description{
Borrowed function from CohortDiagnostics.
}
