% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportFigure.R
\name{exportFigure}
\alias{exportFigure}
\title{exportFigure}
\usage{
exportFigure(
  filePath,
  outputPath,
  zoom = 3,
  vwidth = 430,
  selector = "svg",
  ...
)
}
\arguments{
\item{filePath}{HTML-Filename.}

\item{outputPath}{Filename of image, either with .pdf or .png extension.}

\item{zoom}{Zoom factor, default = 3.}

\item{vwidth}{Width of frame to capture, default = 430.}

\item{selector}{Default: \code{"svg"}; svg only works if the output file is pdf, otherwise use \code{NULL}}

\item{...}{Other potential parameters to be passed to webshot2::webshot()}
}
\value{
Invisibly returns the normalized path to all screenshots taken. The character vector will have a class of '"webshot"'.
}
\description{
Save an SVG-image from a html-file as a PNG, - or PDF-file.
}
\examples{
\dontrun{
  exportFigure("webFile.html", "outputFile.png")}
}
