% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{extractCohortTable}
\alias{extractCohortTable}
\title{extractCohortTable}
\usage{
extractCohortTable(
  connection,
  resultsSchema,
  cohortTableName,
  cohortIds = NULL
)
}
\arguments{
\item{connection}{\link[DatabaseConnector]{connect}\cr
Connection object from DatabaseConnector.}

\item{resultsSchema}{\link[base]{character}\cr
Schema of results.}

\item{cohortTableName}{\link[base]{character}\cr
Name of table.}

\item{cohortIds}{\link[base]{c} of \link[base]{integer}\cr
Vector of cohort ID's.}
}
\value{
\link[base]{data.frame}
Cohort table from database including the specified cohorts\tabular{ll}{
   column \tab data type \cr
   COHORT_DEFINITION_ID \tab \link[base]{integer} \cr
   SUBJECT_ID \tab \link[base]{integer} \cr
   COHORT_START_DATE \tab \link[base]{character} \cr
   COHORT_END_DATE \tab \link[base]{character} \cr
}
}
\description{
Extracts the cohort table from the database, containing specified cohorts.
}
\examples{
\dontrun{
  extractCohortTable(
    connection = connection,
    cohortTableName = "cohortTable",
    resultsSchema = "main",
    cohortIds = c(1,2,3)
  )
}
}
