% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{extractFile}
\alias{extractFile}
\title{extractFile}
\usage{
extractFile(connection, tableName, resultsSchema, dbms)
}
\arguments{
\item{connection}{Connection object}

\item{tableName}{Name of table}

\item{resultsSchema}{Schema of results}

\item{dbms}{DBMS}
}
\value{
data
    data.frame
}
\description{
Extract table with specific name from database server.
}
