% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatSunburstData.R
\name{formatSunburstData}
\alias{formatSunburstData}
\title{formatSunburstData}
\usage{
formatSunburstData(filePath, indexYear)
}
\arguments{
\item{filePath}{\link[base]{character}\cr
Path to file containing `event_cohort_name1`, `event_cohort_name2`,
`event_cohort_name3`, `event_cohort_name4`, `event_cohort_name5` columns.}

\item{indexYear}{\link[base]{integer} or \link[base]{character}\cr
Index year to filter on.}
}
\value{
\link[base]{data.frame}
| column | data type              |
| path   | \link[base]{character} |
| freq   | \link[base]{numeric}   |
}
\description{
Function to format data from files containing event_cohort_name columns.
}
\examples{
\dontrun{
  pathData <- formatSunburstData(
    filePath = "output/treatmentPathways.csv",
    indexYear = NA
  )
  
  createSunburstPlot(
    data = pathData,
    folder = "output/",
    fileName = "sunburst.html"
  )
}
}
