% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateResults.R
\name{generateResults}
\alias{generateResults}
\title{Generate all result files and plots from constructed pathways.}
\usage{
generateResults(databaseName, studyName, outputFolder, tempFolder)
}
\arguments{
\item{databaseName}{Name of the database that will appear in the results.}

\item{studyName}{Name for the study corresponding to the current pathway settings.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes (/).}

\item{tempFolder}{Name of local folder to place intermediate results (not to be shared); make sure to use forward slashes (/).}
}
\description{
Generate all result files and plots from constructed pathways.
}
