% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getCohortCharacteristics}
\alias{getCohortCharacteristics}
\title{getCohortCharacteristics}
\usage{
getCohortCharacteristics(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable = "cohort",
  cohortIds,
  cdmVersion = 5,
  covariateSettings,
  batchSize = 100
)
}
\arguments{
\item{connectionDetails}{Connection details object}

\item{connection}{Connection object}

\item{cdmDatabaseSchema}{CDM database schema}

\item{cohortDatabaseSchema}{Cohort database schema}

\item{cohortTable}{Cohort table}

\item{cohortIds}{Cohort IDs}

\item{cdmVersion}{CDM version}

\item{covariateSettings}{Covariate settings}

\item{batchSize}{Batch size}
}
\description{
Borrowed function from CohortDiagnostics.
}
