% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getCohortCounts}
\alias{getCohortCounts}
\title{getCohortCounts}
\usage{
getCohortCounts(
  connection = NULL,
  cohortDatabaseSchema,
  cohortTable,
  cohortIds = c()
)
}
\arguments{
\item{connection}{Connection object}

\item{cohortDatabaseSchema}{Cohort database schema}

\item{cohortTable}{Cohort table}

\item{cohortIds}{Cohort IDs}
}
\value{
counts
    Number of cohort IDs
}
\description{
Count the cohorts specified by cohortIds (empty = all).
}
