% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getCustomCovariates}
\alias{getCustomCovariates}
\title{getCustomCovariates}
\usage{
getCustomCovariates(
  connection,
  oracleTempSchema = NULL,
  cdmDatabaseSchema,
  cdmVersion = "5",
  cohortTable = "#cohort_person",
  cohortId = -1,
  covariateSettings,
  aggregated = TRUE,
  rowIdField = "subject_id"
)
}
\arguments{
\item{connection}{}

\item{oracleTempSchema}{}

\item{cdmDatabaseSchema}{}

\item{cdmVersion}{}

\item{cohortTable}{}

\item{cohortId}{}

\item{covariateSettings}{}

\item{aggregated}{}

\item{rowIdField}{}
}
\value{
all_covariates
}
\description{
getCustomCovariates
}
\examples{
\dontrun{
connectionDetails <- Eunomia::getEunomiaConnectionDetails()
connection <- DatabaseConnector::connect(connectionDetails)

getCustomCovariates(connection = connection, cdmDatabaseSchema = "main")}
}
