% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOutput.R
\name{getPathways}
\alias{getPathways}
\title{getPathways}
\usage{
getPathways(outputFolder, tempFolder, databaseName, studyName, minCellCount)
}
\arguments{
\item{outputFolder}{Path to the output folder.}

\item{tempFolder}{Path to the temp folder.}

\item{databaseName}{Name of the database that will appear in the results.}

\item{studyName}{Name for the study corresponding to the current settings.}

\item{minCellCount}{Minimum number of persons with a specific treatment pathway for the
pathway to be included in analysis.}
}
\value{
List with two dataframes.
}
\description{
Get treatment pathways for generating aggregate output.
}
