% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResultsDataModelSpecification.R
\name{getResultsDataModelSpecifications}
\alias{getResultsDataModelSpecifications}
\title{getResultsDataModelSpecifications}
\usage{
getResultsDataModelSpecifications()
}
\value{
\code{data.frame}
}
\description{
Gets the results data model specifications of \code{TreatmentPatterns}.
}
\examples{
{
getResultsDataModelSpecification()
}
}
