% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCohorts.R
\name{importCohorts}
\alias{importCohorts}
\title{Only for OMOP-CDM FALSE: Import the target and event cohorts from a csv file in which the cohorts are stored.}
\usage{
importCohorts(cohortLocation, outputFolder)
}
\arguments{
\item{cohortLocation}{Location from where cohorts can be loaded.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes (/).}
}
\description{
Only for OMOP-CDM FALSE: Import the target and event cohorts from a csv file in which the cohorts are stored.
}
