% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOutput.R
\name{inputSunburstPlot}
\alias{inputSunburstPlot}
\title{inputSunburstPlot}
\usage{
inputSunburstPlot(
  data,
  outputFolder,
  databaseName,
  studyName,
  addNoPaths,
  index_year
)
}
\arguments{
\item{data}{Data}

\item{outputFolder}{Output folder}

\item{databaseName}{Database name}

\item{studyName}{Study name}

\item{addNoPaths}{Add no paths}

\item{index_year}{Index year}
}
\value{
transformed_file
}
\description{
Help function to create transform data to correct format for sunburst plot.
}
