% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{is_installed}
\alias{is_installed}
\title{Title}
\usage{
is_installed(pkg, version = 0)
}
\arguments{
\item{pkg}{Package}

\item{version}{Version}
}
\description{
Borrowed from devtools:
https://github.com/hadley/devtools/blob/ba7a5a4abd8258c52cb156e7b26bb4bf47a79f0b/R/utils.r#L44
}
