% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchResultsExplorer.R
\name{launchResultsExplorer}
\alias{launchResultsExplorer}
\title{launchResultsExplorer}
\usage{
launchResultsExplorer(
  saveSettings = NULL,
  zipFolder = NULL,
  outputFolder = NULL
)
}
\arguments{
\item{saveSettings}{Settings object as created by createSaveSettings().}

\item{zipFolder}{Name of local folder containing zip file(s) to move to outputFolder.}

\item{outputFolder}{Name of local folder containing output.}
}
\description{
Processes the zipped result folder(s) and launch the shiny application.
}
