% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{launchShinyApplication}
\alias{launchShinyApplication}
\title{This function processes the zipped result folder(s) and launches the shiny application.}
\usage{
launchShinyApplication(rootFolder)
}
\arguments{
\item{rootFolder}{}
}
\value{

}
\description{
This function processes the zipped result folder(s) and launches the shiny application.
}
