% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPathwaySettings.R
\name{loadPathwaySettings}
\alias{loadPathwaySettings}
\title{loadPathwaySettings}
\usage{
loadPathwaySettings(filePath)
}
\arguments{
\item{filePath}{Path to csv-file containing the pathwaySettings.}
}
\value{
pathwaySettings
}
\description{
Reads in the pathwaySettings from a csv-file.
}
\examples{
loadPathwaySettings(file.path(system.file(
  package = "TreatmentPatterns"),
 "examples", "CDM", "settings", "pathway_settings.csv"))
}
