% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{loadRenderTranslateSql}
\alias{loadRenderTranslateSql}
\title{loadRenderTranslateSql}
\usage{
loadRenderTranslateSql(
  sql,
  oracleTempSchema = NULL,
  dbms = "postgresql",
  warnOnMissingParameters = TRUE,
  output = FALSE,
  outputFile,
  ...
)
}
\arguments{
\item{sql}{SQL statement}

\item{oracleTempSchema}{Temporary Oracle schema}

\item{dbms}{DBMS}

\item{warnOnMissingParameters}{Warning message for missing parameters}

\item{output}{Output}

\item{outputFile}{Output file}

\item{...}{Other parameters}
}
\value{
renderedSql
    renderSql object
}
\description{
Load and translate SQL file to desired dialect.
}
