% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module.R
\name{Module}
\alias{Module}
\title{Module}
\description{
Module super class
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{namespace}}{Namespace of the module.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Module-new}{\code{Module$new()}}
\item \href{#method-Module-validate}{\code{Module$validate()}}
\item \href{#method-Module-uiMenu}{\code{Module$uiMenu()}}
\item \href{#method-Module-uiBody}{\code{Module$uiBody()}}
\item \href{#method-Module-server}{\code{Module$server()}}
\item \href{#method-Module-clone}{\code{Module$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Module-new"></a>}}
\if{latex}{\out{\hypertarget{method-Module-new}{}}}
\subsection{Method \code{new()}}{
Initializer method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Module$new(namespace)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{namespace}}{(\code{character(1)})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{invisible(self)})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Module-validate"></a>}}
\if{latex}{\out{\hypertarget{method-Module-validate}{}}}
\subsection{Method \code{validate()}}{
Validator method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Module$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{invisible(self)})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Module-uiMenu"></a>}}
\if{latex}{\out{\hypertarget{method-Module-uiMenu}{}}}
\subsection{Method \code{uiMenu()}}{
Method to include a \link[shinydashboard]{menuItem} to link to the body.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Module$uiMenu(label, tag)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label}}{(\code{character(1)})\cr
Label to show for the \code{menuItem}.}

\item{\code{tag}}{(\code{character(1)})\cr
Tag to use internally in \code{input}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{menuItem})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Module-uiBody"></a>}}
\if{latex}{\out{\hypertarget{method-Module-uiBody}{}}}
\subsection{Method \code{uiBody()}}{
Method to include a \link[shinydashboard]{tabItem} to include the body.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Module$uiBody()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{tabItem})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Module-server"></a>}}
\if{latex}{\out{\hypertarget{method-Module-server}{}}}
\subsection{Method \code{server()}}{
Method to handle the back-end.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Module$server(input, output, session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{(\code{input})\cr
Input from the server function.}

\item{\code{output}}{(\code{output})\cr
Output from the server function.}

\item{\code{session}}{(\code{session})\cr
Session from the server function.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{NULL})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Module-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Module-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Module$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
