% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOutput.R
\name{outputDurationEras}
\alias{outputDurationEras}
\title{outputDurationEras}
\usage{
outputDurationEras(
  outputFolder,
  tempFolder,
  databaseName,
  studyName,
  eventCohortIds,
  groupCombinations,
  minCellCount
)
}
\arguments{
\item{outputFolder}{Path to the output folder.}

\item{tempFolder}{Path to the temp folder.}

\item{databaseName}{Name of the database that will appear in the results.}

\item{studyName}{Name for the study corresponding to the current settings.}

\item{eventCohortIds}{IDs to refer to event cohorts.}

\item{groupCombinations}{Select to group all non-fixed combinations in one category "other" in
the sunburst plot.}

\item{minCellCount}{Minimum number of persons with a specific treatment pathway for the
pathway to be included in analysis.}
}
\description{
outputDurationEras
}
