% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOutput.R
\name{outputSankeyDiagram}
\alias{outputSankeyDiagram}
\title{outputSankeyDiagram}
\usage{
outputSankeyDiagram(
  data,
  outputFolder,
  databaseName,
  studyName,
  groupCombinations
)
}
\arguments{
\item{data}{Dataframe with event cohorts of the target cohort in different columns.}

\item{outputFolder}{Path to the output folder.}

\item{databaseName}{Name of the database that will appear in the results.}

\item{studyName}{Name for the study corresponding to the current settings.}

\item{groupCombinations}{Select to group all non-fixed combinations in one category "other" in
the sunburst plot.}
}
\description{
outputSankeyDiagram
}
