% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOutput.R
\name{outputTreatedPatients}
\alias{outputTreatedPatients}
\title{outputTreatedPatients}
\usage{
outputTreatedPatients(
  data,
  eventCohortIds,
  groupCombinations,
  outputFolder,
  outputFile
)
}
\arguments{
\item{data}{Dataframe with event cohorts of the target cohort in different columns.}

\item{eventCohortIds}{IDs to refer to event cohorts.}

\item{groupCombinations}{Select to group all non-fixed combinations in one category "other" in
the sunburst plot.}

\item{outputFolder}{Path to the output folder.}

\item{outputFile}{Name of output file.}
}
\description{
outputTreatedPatients
}
