% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOutput.R
\name{percentageGroupTreated}
\alias{percentageGroupTreated}
\title{percentageGroupTreated}
\usage{
percentageGroupTreated(data, eventCohortIds, groupCombinations, outputFolder)
}
\arguments{
\item{data}{Data}

\item{eventCohortIds}{Event cohort IDs}

\item{groupCombinations}{Group combinations}

\item{outputFolder}{Output folder}
}
\value{
result object
}
\description{
Help function to compute percentage of treated patient with certain event
cohorts for outputTreatedPatients.
}
