% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseGaps.R
\name{plotGaps}
\alias{plotGaps}
\title{plotGaps}
\usage{
plotGaps(result, timeScale = "day", ...)
}
\arguments{
\item{result}{(\code{data.frame}) Result from \code{summariseGaps()}}

\item{timeScale}{(\code{character(1)}: \code{"day"}) Timescale to use as the x-axis in
the plot. May be one of: \code{"day"}, \code{"week"}, \code{"month"}, \code{"year"}}

\item{...}{Arguments for \code{visOmopResults::themeVisOmop()}}
}
\value{
\code{ggplot}
}
\description{
Plots the gaps: 1) Between Target cohort entry and the first Event; 2)
Between events; and 3) Between the last Event's end and the Target exit.
}
\examples{
{
if (interactive()) {
  result <- summariseGaps(outputEnv)

  plotGaps(result, timeScale = "week", style = "darwin")
  tableGaps(result)
}
}
}
