% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseOnEvent.R
\name{plotOnEvent}
\alias{plotOnEvent}
\title{plotOnEvent}
\usage{
plotOnEvent(result, ...)
}
\arguments{
\item{result}{(\code{data.frame}) The result of \code{summariseOnEvent()}}

\item{...}{Arguments for \code{visOmopResults::themeVisOmop()}}
}
\value{
\code{ggplot}
}
\description{
Plots the counts of persons that are on an event during the Target cohort per time increment.
}
\examples{
{
if (interactive()) {
  result <- summariseOnEvent(outputEnv, minCellCount = 5)

  plotOnEvent(result)

  # Groups on quarters (90 days)
  result <- summariseOnEvent(outputEnv, minCellCount = 5, timeScale = "quarter" , timeGroup = 90)
  plotOnEvent(result, style = "darwin")
}
}
}
