% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseTargetCohortDecay.R
\name{plotTargetCohortDecay}
\alias{plotTargetCohortDecay}
\title{plotTargetCohortDecay}
\usage{
plotTargetCohortDecay(result, timeScale = "day", ...)
}
\arguments{
\item{result}{(\code{data.frame}) Result from \code{summariseTargetCohortDecay()}}

\item{timeScale}{(\code{character(1)}: \code{"day"}) Timescale to use as the x-axis in the plot. May be one of: \code{"day"}, \code{"week"}, \code{"month"}, \code{"year"}}

\item{...}{Arguments for \code{visOmopResults::themeVisOmop()}}
}
\value{
\code{ggplot}
}
\description{
Plots the decay of the target cohort over time.
}
\examples{
{
if (interactive()) {
  result <- summariseTargetCohortDecay(outputEnv)

  plotTargetCohortDecay(result, timeScale = "week", style = "darwin")
  tableTargetCohortDecay(result)
  tableTargetCohortDecayAtDays(result, timePoints = 1:30, timeScale = "year", style = "darwin")
}
}
}
