% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseTargetIndexToEndOfObservation.R
\name{plotTargetIndexToEndOfObservation}
\alias{plotTargetIndexToEndOfObservation}
\title{plotTargetIndexToEndOfObservation}
\usage{
plotTargetIndexToEndOfObservation(result, ...)
}
\arguments{
\item{result}{(\code{data.frame}) Result from \code{summariseTargetIndexToEndOfObservation()}}

\item{...}{Arguments for \code{visOmopResults::themeVisOmop()}}
}
\value{
\code{ggplot}
}
\description{
Makes a plot out of the result of \code{summariseTargetIndexToEndOfObservation()}.
}
\examples{
{
  if (interactive()) {
    outputEnv <- computePathways(
      cohorts = cohorts,
      cohortTableName = "cohort_table",
      cdm = cdm
    )
    result <- summariseTargetIndexToEndOfObservation(outputEnv, minCellCount = 5)

    plotTargetIndexToEndOfObservation(result, style = "darwin")
  }
}
}
