% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preConfigure.R
\name{preConfigure}
\alias{preConfigure}
\title{preConfigure}
\usage{
preConfigure(saveSettings, cohortSettings, dataSettings, cohortTableNames)
}
\arguments{
\item{saveSettings}{saveSettings object}

\item{cohortSettings}{cohortSettings object}

\item{dataSettings}{dataSettings object}

\item{cohortTableNames}{cohortTableNames}
}
\description{
preConfigure
}
