% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSunburstPlot.R
\name{prepData}
\alias{prepData}
\title{prepData}
\usage{
prepData(treatmentHistory, year)
}
\arguments{
\item{treatmentHistory}{(\code{data.frame()})}

\item{year}{(\code{integer(1)})}
}
\value{
(\code{data.frame()})
}
\description{
prepData
}
