% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOutput.R
\name{preprocessSunburstData}
\alias{preprocessSunburstData}
\title{preprocessSunburstData}
\usage{
preprocessSunburstData(
  data,
  outputFolder,
  databaseName,
  studyName,
  eventCohortIds,
  addNoPaths
)
}
\arguments{
\item{data}{Dataframe with event cohorts of the target cohort in different columns.}

\item{outputFolder}{Path to the output folder.}

\item{databaseName}{Name of the database that will appear in the results.}

\item{studyName}{Name for the study corresponding to the current settings.}

\item{eventCohortIds}{IDs to refer to event cohorts.}

\item{addNoPaths}{Select to include untreated persons without treatment pathway in the
sunburst plot.}
}
\description{
Preprocesses the data for the sunburst either total or by year
}
