% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSettings.R
\name{readInPathwaySettings}
\alias{readInPathwaySettings}
\title{readInPathwaySettings}
\usage{
readInPathwaySettings(filePath)
}
\arguments{
\item{filePath}{Path to csv-file containing the pathwaySettings.}
}
\description{
Reads in the pathwaySettings from a csv-file.
}
\examples{
readInPathwaySettings(file.path(system.file(
  package = "TreatmentPatterns"),
 "examples", "OMOP CDM", "inst", "settings", "pathway_settings.csv"))
}
