% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SankeyDiagram.R
\name{SankeyDiagram}
\alias{SankeyDiagram}
\title{SankeyDiagram}
\description{
Class to handle the Sankey diagram of TreatmentPatterns.
}
\section{Super classes}{
\code{\link[TreatmentPatterns:ShinyModule]{TreatmentPatterns::ShinyModule}} -> \code{TreatmentPatterns::InteracitvePlot} -> \code{SankeyDiagram}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SankeyDiagram-clone}{\code{SankeyDiagram$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="TreatmentPatterns" data-topic="ShinyModule" data-id="validate"><a href='../../TreatmentPatterns/html/ShinyModule.html#method-ShinyModule-validate'><code>TreatmentPatterns::ShinyModule$validate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TreatmentPatterns" data-topic="InteracitvePlot" data-id="initialize"><a href='../../TreatmentPatterns/html/InteracitvePlot.html#method-InteracitvePlot-initialize'><code>TreatmentPatterns::InteracitvePlot$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TreatmentPatterns" data-topic="InteracitvePlot" data-id="server"><a href='../../TreatmentPatterns/html/InteracitvePlot.html#method-InteracitvePlot-server'><code>TreatmentPatterns::InteracitvePlot$server()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TreatmentPatterns" data-topic="InteracitvePlot" data-id="uiBody"><a href='../../TreatmentPatterns/html/InteracitvePlot.html#method-InteracitvePlot-uiBody'><code>TreatmentPatterns::InteracitvePlot$uiBody()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TreatmentPatterns" data-topic="InteracitvePlot" data-id="uiMenu"><a href='../../TreatmentPatterns/html/InteracitvePlot.html#method-InteracitvePlot-uiMenu'><code>TreatmentPatterns::InteracitvePlot$uiMenu()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SankeyDiagram-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SankeyDiagram-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SankeyDiagram$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
