% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveAsFile.R
\name{saveAsFile}
\alias{saveAsFile}
\title{saveAsFile}
\usage{
saveAsFile(fileName, fileNameOut, zoom = 3, vwidth = 430, ...)
}
\arguments{
\item{fileName}{HTML-Filename.}

\item{fileNameOut}{Filename of image, either with .pdf or .png extension.}

\item{zoom}{Zoom factor, default = 3.}

\item{vwidth}{Width of frame to capture, default = 430.}

\item{...}{Other potential parameters to be passed to webshot2::webshot()}
}
\description{
saveAsFile
}
\examples{
\dontrun{
  saveAsFile("webFile.html", "outputFile.png")}
}
