% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOutput.R
\name{saveAsPNG}
\alias{saveAsPNG}
\title{saveAsPNG}
\usage{
saveAsPNG(fileName, fileNameOut, zoom = 3, vwidth = 430, ...)
}
\arguments{
\item{fileName}{File name of the input HTML-file.}

\item{fileNameOut}{Filename of the output PNG-file.}

\item{zoom}{Zoom factor}

\item{vwidth}{Width of the image}

\item{...}{Other parameters that webshot2::webshot can use. See ??webshot2::websot}
}
\description{
Takes a scaled screenshot of a html page. Shadows webshot2::webshot.
}
