% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-SaveSettings.R, R/SaveSettings.R
\name{SaveSettings}
\alias{SaveSettings}
\title{SaveSettings}
\description{
R6 SaveSettings object.
}
\details{
SaveSettings R6 object. Inherits from Settings.
}
\examples{
saveSettings <- SaveSettings$new(
  databaseName = "Eunomia",
  outputFolder = "output"
)
}
\seealso{
Other Settings: 
\code{\link{AnalysisSettings}},
\code{\link{CohortSettings}},
\code{\link{DataSettings}},
\code{\link{PathwaySettings}}
}
\concept{Settings}
\section{Super class}{
\code{TreatmentPatterns::Settings} -> \code{SaveSettings}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SaveSettings-new}{\code{SaveSettings$new()}}
\item \href{#method-SaveSettings-get}{\code{SaveSettings$get()}}
\item \href{#method-SaveSettings-validate}{\code{SaveSettings$validate()}}
\item \href{#method-SaveSettings-clone}{\code{SaveSettings$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="TreatmentPatterns" data-topic="Settings" data-id="print"><a href='../../TreatmentPatterns/html/Settings.html#method-Settings-print'><code>TreatmentPatterns::Settings$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaveSettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-SaveSettings-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveSettings$new(databaseName, outputFolder)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaveSettings-get"></a>}}
\if{latex}{\out{\hypertarget{method-SaveSettings-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveSettings$get()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaveSettings-validate"></a>}}
\if{latex}{\out{\hypertarget{method-SaveSettings-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveSettings$validate(databaseName, outputFolder)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaveSettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SaveSettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{TreatmentPatterns::Settings} -> \code{SaveSettings}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SaveSettings-new}{\code{SaveSettings$new()}}
\item \href{#method-SaveSettings-get}{\code{SaveSettings$get()}}
\item \href{#method-SaveSettings-validate}{\code{SaveSettings$validate()}}
\item \href{#method-SaveSettings-clone}{\code{SaveSettings$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="TreatmentPatterns" data-topic="Settings" data-id="print"><a href='../../TreatmentPatterns/html/Settings.html#method-Settings-print'><code>TreatmentPatterns::Settings$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaveSettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-SaveSettings-new}{}}}
\subsection{Method \code{new()}}{
Initialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveSettings$new(databaseName, outputFolder)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{databaseName}}{\if{html}{\out{<character>}} Database name.}

\item{\code{outputFolder}}{\if{html}{\out{<character>}} Path to output folder.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(self)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaveSettings-get"></a>}}
\if{latex}{\out{\hypertarget{method-SaveSettings-get}{}}}
\subsection{Method \code{get()}}{
get
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveSettings$get()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\if{html}{\out{
  <list>
}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaveSettings-validate"></a>}}
\if{latex}{\out{\hypertarget{method-SaveSettings-validate}{}}}
\subsection{Method \code{validate()}}{
validate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveSettings$validate(databaseName, outputFolder)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{databaseName}}{\if{html}{\out{<character>}} Database name.}

\item{\code{outputFolder}}{\if{html}{\out{<character>}} Path to output folder.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(self)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaveSettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SaveSettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
