% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructPathways.R
\name{selectRowsCombinationWindow}
\alias{selectRowsCombinationWindow}
\title{selectRowsCombinationWindow}
\usage{
selectRowsCombinationWindow(treatment_history)
}
\arguments{
\item{treatment_history}{Dataframe with event cohorts of the target cohort in different rows.}
}
\value{
Updated treatment_history data.frame
}
\description{
Help function for doCombinationWindow that selects one overlapping drug era
per person to modify in next iteration of the combination window.
}
\examples{
\dontrun{
source(system.file(
  package = "TreatmentPatterns", 
 "testing", "testParams.R"))
 
selectRowsCombinationWindow(doEraCollapseTH)
}
}
