% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCohortTypes.R
\name{setCohortTypes}
\alias{setCohortTypes}
\title{setCohortTypes}
\usage{
setCohortTypes(x, types)
}
\arguments{
\item{x}{(\code{data.frame}) Either a \code{data.frame} that is filled with records
from \code{CohortGenerator::createEmptyCohortDefinitionSet()} or
\code{CohortGenerator::generateCohortSet()}; or a \code{CohortSet} generated by
\code{CDMConnector::readCohortSet()}}

\item{types}{(\code{character(n)}) A character vector with the types. Must
contain a selection of: \code{"target"}, \code{"event"}, or \code{"exit"}. The length of
the vector must be equal to the number of rows in \code{x}}
}
\value{
\code{data.frame}
}
\description{
Creates a data.frame to be used as the \code{cohorts} argument in \code{computePathways()}.
}
\examples{
{
if (interactive()) {
  cohorts <- setCohortTypes(cohortSet, types = c(rep("event", 7), "target"))
  cohorts <- setCohortTypes(cohortsGenerated, types = c("target", rep("event", 7)))
  cohorts <- setCohortTypes(cohortsToCreate, types = c("target", rep("event", 7)))
}
}
}
