% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\name{Settings}
\alias{Settings}
\title{Settings}
\description{
Settings

Settings
}
\details{
Settings super object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Settings-new}{\code{Settings$new()}}
\item \href{#method-Settings-get}{\code{Settings$get()}}
\item \href{#method-Settings-print}{\code{Settings$print()}}
\item \href{#method-Settings-validate}{\code{Settings$validate()}}
\item \href{#method-Settings-clone}{\code{Settings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Settings-new"></a>}}
\if{latex}{\out{\hypertarget{method-Settings-new}{}}}
\subsection{Method \code{new()}}{
Initialize method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Settings$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible(self)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Settings-get"></a>}}
\if{latex}{\out{\hypertarget{method-Settings-get}{}}}
\subsection{Method \code{get()}}{
get method to overload
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Settings$get()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Settings-print"></a>}}
\if{latex}{\out{\hypertarget{method-Settings-print}{}}}
\subsection{Method \code{print()}}{
print method to overload
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Settings$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible(self)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Settings-validate"></a>}}
\if{latex}{\out{\hypertarget{method-Settings-validate}{}}}
\subsection{Method \code{validate()}}{
validate method to overload
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Settings$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible(self)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Settings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Settings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Settings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
