% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyModule.R
\name{ShinyModule}
\alias{ShinyModule}
\title{ShinyModule}
\description{
ShinyModule super class
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{namespace}}{Namespace of the module.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ShinyModule-new}{\code{ShinyModule$new()}}
\item \href{#method-ShinyModule-validate}{\code{ShinyModule$validate()}}
\item \href{#method-ShinyModule-uiMenu}{\code{ShinyModule$uiMenu()}}
\item \href{#method-ShinyModule-uiBody}{\code{ShinyModule$uiBody()}}
\item \href{#method-ShinyModule-server}{\code{ShinyModule$server()}}
\item \href{#method-ShinyModule-clone}{\code{ShinyModule$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShinyModule-new"></a>}}
\if{latex}{\out{\hypertarget{method-ShinyModule-new}{}}}
\subsection{Method \code{new()}}{
Initializer method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyModule$new(namespace)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{namespace}}{(\code{character(1)})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{invisible(self)})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShinyModule-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ShinyModule-validate}{}}}
\subsection{Method \code{validate()}}{
Validator method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyModule$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{invisible(self)})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShinyModule-uiMenu"></a>}}
\if{latex}{\out{\hypertarget{method-ShinyModule-uiMenu}{}}}
\subsection{Method \code{uiMenu()}}{
Method to include a \link[shinydashboard]{menuItem} to link to the body.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyModule$uiMenu(label, tag)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label}}{(\code{character(1)})\cr
Label to show for the \code{menuItem}.}

\item{\code{tag}}{(\code{character(1)})\cr
Tag to use internally in \code{input}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{menuItem})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShinyModule-uiBody"></a>}}
\if{latex}{\out{\hypertarget{method-ShinyModule-uiBody}{}}}
\subsection{Method \code{uiBody()}}{
Method to include a \link[shinydashboard]{tabItem} to include the body.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyModule$uiBody()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{tabItem})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShinyModule-server"></a>}}
\if{latex}{\out{\hypertarget{method-ShinyModule-server}{}}}
\subsection{Method \code{server()}}{
Method to handle the back-end.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyModule$server(input, output, session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{(\code{input})\cr
Input from the server function.}

\item{\code{output}}{(\code{output})\cr
Output from the server function.}

\item{\code{session}}{(\code{session})\cr
Session from the server function.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{NULL})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShinyModule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ShinyModule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyModule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
