% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{stratisfy}
\alias{stratisfy}
\title{stratisfy}
\usage{
stratisfy(treatmentHistory, years, ages)
}
\arguments{
\item{treatmentHistory}{(\code{data.frame()})
Patient level Treatment History data.frame}

\item{years}{(\code{vector("character")})}

\item{ages}{(\code{vector("character")})}
}
\value{
(\code{data.frame()})
}
\description{
stratisfy
}
