% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{stripname}
\alias{stripname}
\title{stripname}
\usage{
stripname(x, name)
}
\arguments{
\item{x}{List}

\item{name}{Name}
}
\value{
List with removed names
}
\description{
Recursive function to remove name from all levels of list.
}
