% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseGaps.R
\name{summariseGaps}
\alias{summariseGaps}
\title{summariseGaps}
\usage{
summariseGaps(andromeda)
}
\arguments{
\item{andromeda}{(\code{Andromeda}) Andromeda object returned by \code{computePathways()}}
}
\value{
\code{data.frame}
}
\description{
Summarises the gaps: 1) Between Target cohort entry and the first Event; 2)
Between events; and 3) Between the last Event's end and the Target exit.
}
\examples{
{
if (interactive()) {
  result <- summariseGaps(outputEnv)

  plotGaps(result, timeScale = "week", style = "darwin")
  tableGaps(result)
}
}
}
