% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseOnEvent.R
\name{summariseOnEvent}
\alias{summariseOnEvent}
\title{summariseOnEvent}
\usage{
summariseOnEvent(andromeda, minCellCount, timeScale = "day", timeGroup = NULL)
}
\arguments{
\item{andromeda}{(\code{Andromeda}) Andromeda object returned by \code{computePathways()}}

\item{minCellCount}{(\code{numeric(1)}) Minimum cell count. Censors any count below it to \code{NA}.}

\item{timeScale}{(\code{character(1)}: \code{"day"}) Sets the time scale to group the counts on. Also used as a label when \code{timeGroup} is set.}

\item{timeGroup}{(\code{numeric(1)}: \code{NULL}) Sets the time group associated with the \code{timeScale}.}
}
\value{
\code{data.frame}
}
\description{
Summarises the counts of persons that are on an event during the Target cohort per time increment.
}
\examples{
{
if (interactive()) {
  result <- summariseOnEvent(outputEnv, minCellCount = 5)

  # Groups on quarters (90 days)
  result <- summariseOnEvent(outputEnv, minCellCount = 5, timeScale = "quarter" , timeGroup = 90)
}
}
}
