% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseTargetCohortDecay.R
\name{summariseTargetCohortDecay}
\alias{summariseTargetCohortDecay}
\title{summariseTargetCohortDecay}
\usage{
summariseTargetCohortDecay(andromeda, minCellCount)
}
\arguments{
\item{andromeda}{(\code{Andromeda}) Andromeda object returned by \code{computePathways()}}

\item{minCellCount}{(\code{numeric(1)}) Cut off value to censor any counts, that fall below this value.}
}
\value{
\code{data.frame}
}
\description{
Summarises the decay of the Target cohort over time.
}
\examples{
{
if (interactive()) {
  result <- summariseTargetCohortDecay(outputEnv)

  plotTargetCohortDecay(result, timeScale = "week", style = "darwin")
  tableTargetCohortDecay(result)
  tableTargetCohortDecayAtDays(result, timePoints = 1:30, timeScale = "year", style = "darwin")
}
}
}
