% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseTargetIndexToEndOfObservation.R
\name{summariseTargetIndexToEndOfObservation}
\alias{summariseTargetIndexToEndOfObservation}
\title{summariseTargetIndexToEndOfObservation}
\usage{
summariseTargetIndexToEndOfObservation(andromeda, minCellCount)
}
\arguments{
\item{andromeda}{(\code{Andromeda}) Andromeda object returned by \code{computePathways()}}

\item{minCellCount}{(\code{numeric(1)}) Any count values will be cencored below this value.}
}
\value{
\code{data.frame}
}
\description{
Computes summary statistics of the duration from the index date of the
Target cohort to the end of observation.
}
\examples{
{
  if (interactive()) {
    outputEnv <- computePathways(
      cohorts = cohorts,
      cohortTableName = "cohort_table",
      cdm = cdm
    )
    result <- summariseTargetIndexToEndOfObservation(outputEnv, minCellCount = 5)
  }
}
}
