% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseGaps.R
\name{tableGaps}
\alias{tableGaps}
\title{tableGaps}
\usage{
tableGaps(result, timeScale = "day", ...)
}
\arguments{
\item{result}{(\code{data.frame}) Result from \code{summariseGaps()}}

\item{timeScale}{(\code{character(1)}: \code{"day"}) Time scale to use for the summary
statistics. May be one of: \code{"day"}, \code{"week"}, \code{"month"}, \code{"year"}}

\item{...}{Arguments for \code{visOmopResults::visTable()}}
}
\value{
\code{data.frame}
}
\description{
Creates a table of summary statistics of the gaps: 1) Between Target cohort
entry and the first Event; 2) Between events; and 3) Between the last
Event's end and the Target exit.
}
\examples{
{
if (interactive()) {
  result <- summariseGaps(outputEnv)

  plotGaps(result, timeScale = "week", style = "darwin")
  tableGaps(result)
}
}
}
