% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseOnEvent.R
\name{tableOnEvent}
\alias{tableOnEvent}
\title{tableOnEvent}
\usage{
tableOnEvent(result, .force = FALSE, ...)
}
\arguments{
\item{result}{(\code{data.frame}) The result of \code{summariseOnEvent()}}

\item{.force}{(\code{logical(1)}) A check is implemented to make sure you really want to return a table of > 1000 rows (usually no). \code{.force} circumvents this check.}

\item{...}{Arguments for \code{visOmopResults::visTable()}}
}
\value{
A formatted table of the class selected in "type" argument.
}
\description{
Tables the counts of persons that are on an event during the Target cohort per time increment.
}
\examples{
{
if (interactive()) {
  result <- summariseOnEvent(outputEnv, minCellCount = 5)

  tableOnEvent(result)

  # Groups on quarters (90 days)
  result <- summariseOnEvent(outputEnv, minCellCount = 5, timeScale = "quarter" , timeGroup = 90)
  tableOnEvent(result, style = "darwin")
}
}
}
