% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseTargetCohortDecay.R
\name{tableTargetCohortDecay}
\alias{tableTargetCohortDecay}
\title{tableTargetCohortDecay}
\usage{
tableTargetCohortDecay(result, timeScale = "day", ...)
}
\arguments{
\item{result}{(\code{data.frame}) Result from \code{summariseTargetCohortDecay()}}

\item{timeScale}{(\code{character(1)}: \code{"day"}) Time scale to use for the summary statistics. May be one of: \code{"day"}, \code{"week"}, \code{"month"}, \code{"year"}}

\item{...}{Arguments for \code{visOmopResults::visTable()}}
}
\value{
\code{data.frame}
}
\description{
Creates a table of summary statistics of the
}
\examples{
{
if (interactive()) {
  result <- summariseTargetCohortDecay(outputEnv)

  plotTargetCohortDecay(result, timeScale = "week", style = "darwin")
  tableTargetCohortDecay(result)
  tableTargetCohortDecayAtDays(result, timePoints = 1:30, timeScale = "year", style = "darwin")
}
}
}
