% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseTargetCohortDecay.R
\name{tableTargetCohortDecayAtDays}
\alias{tableTargetCohortDecayAtDays}
\title{tableTargetCohortDecayAtDays}
\usage{
tableTargetCohortDecayAtDays(result, timePoints, timeScale = "day", ...)
}
\arguments{
\item{result}{(\code{data.frame}) Result from \code{summariseTargetCohortDecay()}}

\item{timePoints}{(\code{numeric(n)}) A vector of time points.}

\item{timeScale}{(\code{character(1)}: \code{"day"}) Time scale to use for the summary statistics. May be one of: \code{"day"}, \code{"week"}, \code{"month"}, \code{"year"}}

\item{...}{Arguments for \code{visOmopResults::visTable()}}
}
\value{
\code{data.frame}
}
\description{
Creates a table of the number of subjects in the Target cohort, at given days, weeks, months or years.
}
\examples{
{
if (interactive()) {
  result <- summariseTargetCohortDecay(outputEnv)

  plotTargetCohortDecay(result, timeScale = "week", style = "darwin")
  tableTargetCohortDecay(result)
  tableTargetCohortDecayAtDays(result, timePoints = 1:30, timeScale = "year", style = "darwin")
}
}
}
