% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseTargetIndexToEndOfObservation.R
\name{tableTargetIndexToEndOfObservation}
\alias{tableTargetIndexToEndOfObservation}
\title{tableTargetIndexToEndOfObservation}
\usage{
tableTargetIndexToEndOfObservation(result, ...)
}
\arguments{
\item{result}{(\code{data.frame}) Result from \code{summariseTargetIndexToEndOfObservation()}}

\item{...}{Arguments for \code{visOmopResults::visTable()}}
}
\value{
A formatted table of the class selected in "type" argument.
}
\description{
Makes a table out of the result of \code{summariseTargetIndexToEndOfObservation()} using \code{visOmopResults::visTable()}
}
\examples{
{
  if (interactive()) {
    outputEnv <- computePathways(
      cohorts = cohorts,
      cohortTableName = "cohort_table",
      cdm = cdm
    )
    result <- summariseTargetIndexToEndOfObservation(outputEnv, minCellCount = 5)

    tableTargetIndexToEndOfObservation(result, style = "darwin")
  }
}
}
