% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableTreatmentPathways.R
\name{tableTreatmentPathways}
\alias{tableTreatmentPathways}
\title{tableTreatmentPathways}
\usage{
tableTreatmentPathways(result, ...)
}
\arguments{
\item{result}{(\code{TreatmentPatternsResults}) Result object from \code{TreatmentPatterns::export()}.}

\item{...}{Additional arguments are passed to \code{visOmopResults::visTable()}.}
}
\value{
A formatted table of the class selected in "type" argument.
}
\description{
Function that creates a formatted table of the Treatment Pathways result using \code{visOmopResults}.
}
\examples{
{
if (interactive()) {
  tableTreatmentPathways(result, style = "darwin")
}
}
}
