% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSunburstPlot.R
\name{toFile}
\alias{toFile}
\title{toFile}
\usage{
toFile(json, treatmentPathways, outputFile)
}
\arguments{
\item{json}{(\code{character(1)})}

\item{treatmentPathways}{(\code{data.frame()})}

\item{outputFile}{(\code{character(1)})}
}
\value{
(\code{NULL})
}
\description{
toFile
}
