% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPSettings.R
\name{TPSettings}
\alias{TPSettings}
\title{TPSettings}
\description{
TPSettings

TPSettings
}
\details{
TPSettings object, inherits from Settings.
}
\section{Super class}{
\code{TreatmentPatterns::Settings} -> \code{TPSettings}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TPSettings-new}{\code{TPSettings$new()}}
\item \href{#method-TPSettings-clone}{\code{TPSettings$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="TreatmentPatterns" data-topic="Settings" data-id="get"><a href='../../TreatmentPatterns/html/Settings.html#method-Settings-get'><code>TreatmentPatterns::Settings$get()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TreatmentPatterns" data-topic="Settings" data-id="print"><a href='../../TreatmentPatterns/html/Settings.html#method-Settings-print'><code>TreatmentPatterns::Settings$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TreatmentPatterns" data-topic="Settings" data-id="validate"><a href='../../TreatmentPatterns/html/Settings.html#method-Settings-validate'><code>TreatmentPatterns::Settings$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPSettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-TPSettings-new}{}}}
\subsection{Method \code{new()}}{
initialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPSettings$new(dataSettings, cohortSettings, pathwaySettings, saveSettings)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataSettings}}{<DataSettings> DataSettings R6 object.}

\item{\code{cohortSettings}}{<CohortSettings> CohortSettings R6 object.}

\item{\code{pathwaySettings}}{<PathwaySettings> PathwaySettings R6 object.}

\item{\code{saveSettings}}{<SaveSettings> SaveSettings R6 object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(self)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPSettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TPSettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
