% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOutput.R
\name{transformCSVtoJSON}
\alias{transformCSVtoJSON}
\title{transformCSVtoJSON}
\usage{
transformCSVtoJSON(data, outcomes, folder, file_name)
}
\arguments{
\item{data}{input data.frame}

\item{outcomes}{Outcomes}

\item{folder}{output folder}

\item{file_name}{output file name}
}
\value{
JSON

result
}
\description{
Help function to transform data in csv format to required JSON format for
HTML.
}
\examples{
\dontrun{
transformCSVtoJSON(data = data.frame(path = "", freq = 1),
                   outcomes = 1,
                   folder = getwd(),
                   file_name = "result")
}
}
