% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatmentPatterns-package.R
\docType{package}
\name{TreatmentPatterns-package}
\alias{TreatmentPatterns}
\alias{TreatmentPatterns-package}
\title{TreatmentPatterns: Analyzes Real-World Treatment Patterns of a Study Population of Interest}
\description{
Computes treatment patterns within a given cohort using the Observational Medical Outcomes Partnership (OMOP) commen data model (CDM).
}
\keyword{internal}
