% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchResultsExplorer.R
\name{unzipFiles}
\alias{unzipFiles}
\title{unzipFiles}
\usage{
unzipFiles(rootFolder, unzipMainFolder)
}
\arguments{
\item{rootFolder}{Root folder}

\item{unzipMainFolder}{Unzip main folder}
}
\description{
Borrowed and adapted function from CohortDiagnostics.
}
