% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{writeCohortTable}
\alias{writeCohortTable}
\title{writeCohortTable}
\usage{
writeCohortTable(saveSettings, cohortSettings, dataSettings, tableName)
}
\arguments{
\item{saveSettings}{saveSettings object}

\item{cohortSettings}{cohortSettings object}

\item{dataSettings}{dataSettings object}

\item{tableName}{Name of the cohort table in the database.}
}
\description{
Writes the cohortTable from the database to a specified path in saveSettings.
}
