% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{writeCovariateDataAndromedaToCsv}
\alias{writeCovariateDataAndromedaToCsv}
\title{writeCovariateDataAndromedaToCsv}
\usage{
writeCovariateDataAndromedaToCsv(data, fileName, incremental = FALSE)
}
\arguments{
\item{data}{Data}

\item{fileName}{Filename}

\item{incremental}{Incremental}
}
\description{
Borrowed function from CohortDiagnostics.
}
