% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{writeToCsv}
\alias{writeToCsv}
\title{writeToCsv}
\usage{
writeToCsv(data, fileName, incremental = FALSE, ...)
}
\arguments{
\item{data}{Data}

\item{fileName}{File name}

\item{incremental}{Incremental}

\item{...}{Other parameters}
}
\description{
Borrowed function from CohortDiagnostics.
}
