% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructPathways.R
\name{addLabels}
\alias{addLabels}
\title{addLabels}
\usage{
addLabels(treatment_history, outputFolder)
}
\arguments{
\item{treatment_history}{treatment_history object}

\item{outputFolder}{Folder of output}
}
\value{
treatment_history
}
\description{
Adds back cohort names to concept ids.
}
