% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructPathways.R
\name{addLabels}
\alias{addLabels}
\title{addLabels}
\usage{
addLabels(andromeda)
}
\arguments{
\item{andromeda}{(\code{Andromeda::andromeda()})}
}
\value{
(invisible(NULL))
}
\description{
Adds back cohort names to concept ids.
}
