% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPathwayAnalysis.R
\name{addPathwayAnalysis}
\alias{addPathwayAnalysis}
\title{addPathwayAnalysis}
\usage{
addPathwayAnalysis(pathwaySettings, targetCohortIds, eventCohortIds, ...)
}
\arguments{
\item{pathwaySettings}{A pathwaySettings object}

\item{targetCohortIds}{Target Cohort ID's}

\item{eventCohortIds}{Event Cohort ID's}

\item{...}{Optional addPathwaySettings parameters i.e. studyName. Excluding
eventCohortId and targetCohortIds}
}
\value{
pathwaySettings
}
\description{
Adds a pathway analysis, a combination of target and outcome cohorts, to a pathwaySettings object.
}
\examples{
# Create cohortSettings
cohortSettings <- createCohortSettings(
  targetCohorts = data.frame(
    cohortId = c(1),
    cohortName = c("a")
  ),
  eventCohorts = data.frame(
    cohortId = c(2, 3),
    cohortName = c("b", "c")
  )
)

# Create pathwaySettings
pathwaySettings <- createPathwaySettings(cohortSettings)

addPathwayAnalysis(
  pathwaySettings = pathwaySettings,
  targetCohortIds = c(10),
  eventCohortIds = c(11, 12),
  studyName = "Second study"
)
}
