% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPathwaySettings.R
\name{addPathwaySettings}
\alias{addPathwaySettings}
\title{addPathwaySettings}
\usage{
addPathwaySettings(
  studyName = "nameUnknown",
  targetCohortId,
  eventCohortIds,
  exitCohortIds = NULL,
  includeTreatments = "startDate",
  periodPriorToIndex = 0,
  minEraDuration = 0,
  splitEventCohorts = "",
  splitTime = 30,
  eraCollapseSize = 30,
  combinationWindow = 30,
  minPostCombinationDuration = 30,
  filterTreatments = "First",
  maxPathLength = 5,
  minCellCount = 5,
  minCellMethod = "Remove",
  groupCombinations = 10,
  addNoPaths = TRUE
)
}
\arguments{
\item{studyName}{(\link[base]{character}: "nameUnknown")\cr
Name identifying the set of study parameters.}

\item{targetCohortId}{(\link[base]{c}) of (\link[base]{numeric})\cr
Vector of target cohort ID of current study settings.}

\item{eventCohortIds}{(\link[base]{c}) of (\link[base]{numeric})\cr
Vector event cohort IDs of current study settings.}

\item{exitCohortIds}{(\link[base]{c}) of (\link[base]{numeric})\cr
Vector exit cohort IDs of current study settings.}

\item{includeTreatments}{(\link[base]{character}: "startDate") ["startDate", "endDate"]\cr
Include treatments starting "startDate" or ending "endDate" after target
cohort start date.}

\item{periodPriorToIndex}{(\link[base]{numeric}: 0)\cr
Number of days prior to the index date of the
target cohort that event cohorts are allowed to start.}

\item{minEraDuration}{(\link[base]{numeric}: 0)\cr
Minimum time an event era should last to be
included in analysis.}

\item{splitEventCohorts}{(\link[base]{character}: "")\cr
Specify event cohort ID's to split in acute
(< X days) and therapy (>= X days).}

\item{splitTime}{(\link[base]{numeric}: 30)\cr
Specify number of days (X) at which each of the split event cohorts should
be split in acute and therapy.}

\item{eraCollapseSize}{(\link[base]{numeric}: 30)\cr
Window of time between which two eras of the same event cohort are collapsed
into one era.}

\item{combinationWindow}{(\link[base]{numeric}: 30)\cr
Window of time two event cohorts need to overlap to be considered a
combination treatment.}

\item{minPostCombinationDuration}{(\link[base]{numeric}: 30)\cr
Minimum time an event era before or after a generated combination treatment
should last to be included in analysis.}

\item{filterTreatments}{(\link[base]{character}: "First") ["First", "Changes", "All"
]\cr
Select first occurrence of ("First"); changes between ("Changes'); all event
cohorts ("All").}

\item{maxPathLength}{(\link[base]{numeric}: 5)\cr
Maximum number of steps included in treatment pathway. Up to 5 is supported.}

\item{minCellCount}{(\link[base]{numeric}: 5)\cr
Minimum number of persons with a specific treatment pathway for the pathway
to be included in analysis.}

\item{minCellMethod}{(\link[base]{character}: "Remove") ["Remove"]\cr
Select to completely remove / sequentially adjust (by removing last step
as often as necessary) treatment pathways below minCellCount.}

\item{groupCombinations}{(\link[base]{numeric}: 10)\cr
Select to group all non-fixed combinations in one category 'other’ in the
sunburst plot.}

\item{addNoPaths}{(\link[base]{logical}: TRUE)\cr
Select to include untreated persons without treatment pathway in the
sunburst plot. Default value is FALSE. When set to TRUE the sunburst plot
will show non-treated individuals as blank space.}
}
\value{
(\link[base]{data.frame})\cr
A data frame containing the pathway settings.
}
\description{
Defines and returns a data.frame specifying different parameters how to
compute the existing pathways in the function call.
}
\examples{
\dontrun{
pathwaySettings <- addPathwaySettings(
  studyName = "myCoolStudy",
  targetCohortId = 1,
  eventCohortIds = c(1,2,3))
}
}
