% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPathwaySettings.R
\name{addPathwaySettingsCheck}
\alias{addPathwaySettingsCheck}
\title{addPathwaySettingsCheck}
\usage{
addPathwaySettingsCheck(
  studyName,
  targetCohortId,
  eventCohortIds,
  includeTreatments,
  periodPriorToIndex,
  minEraDuration,
  splitEventCohorts,
  splitTime,
  eraCollapseSize,
  combinationWindow,
  minPostCombinationDuration,
  filterTreatments,
  maxPathLength,
  minCellCount,
  minCellMethod,
  groupCombinations,
  addNoPaths
)
}
\arguments{
\item{studyName}{}

\item{targetCohortId}{}

\item{eventCohortIds}{}

\item{includeTreatments}{}

\item{periodPriorToIndex}{}

\item{minEraDuration}{}

\item{splitEventCohorts}{}

\item{splitTime}{}

\item{eraCollapseSize}{}

\item{combinationWindow}{}

\item{minPostCombinationDuration}{}

\item{filterTreatments}{}

\item{maxPathLength}{}

\item{minCellCount}{}

\item{minCellMethod}{}

\item{groupCombinations}{}

\item{addNoPaths}{}
}
\value{
TRUE if all assertions pass
}
\description{
addPathwaySettingsCheck
}
