% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-AnalysisSettings.R
\name{AnalysisSettings}
\alias{AnalysisSettings}
\title{AnalysisSettings}
\description{
AnalysisSettings

AnalysisSettings
}
\section{Super class}{
\code{\link[TreatmentPatterns:Settings]{TreatmentPatterns::Settings}} -> \code{AnalysisSettings}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnalysisSettings-new}{\code{AnalysisSettings$new()}}
\item \href{#method-AnalysisSettings-editSettings}{\code{AnalysisSettings$editSettings()}}
\item \href{#method-AnalysisSettings-get}{\code{AnalysisSettings$get()}}
\item \href{#method-AnalysisSettings-validate}{\code{AnalysisSettings$validate()}}
\item \href{#method-AnalysisSettings-clone}{\code{AnalysisSettings$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="TreatmentPatterns" data-topic="Settings" data-id="print"><a href='../../TreatmentPatterns/html/Settings.html#method-Settings-print'><code>TreatmentPatterns::Settings$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnalysisSettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnalysisSettings-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalysisSettings$new(cohortSettings, studyName)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnalysisSettings-editSettings"></a>}}
\if{latex}{\out{\hypertarget{method-AnalysisSettings-editSettings}{}}}
\subsection{Method \code{editSettings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalysisSettings$editSettings(
  includeTreatments = "startDate",
  periodPriorToIndex = 0,
  minEraDuration = 0,
  splitEventCohorts = "",
  splitTime = 30,
  eraCollapseSize = 30,
  combinationWindow = 30,
  minPostCombinationDuration = 30,
  filterTreatments = "First",
  maxPathLength = 5,
  minCellCount = 5,
  minCellMethod = "Remove",
  groupCombinations = 10,
  addNoPaths = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnalysisSettings-get"></a>}}
\if{latex}{\out{\hypertarget{method-AnalysisSettings-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalysisSettings$get()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnalysisSettings-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AnalysisSettings-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalysisSettings$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnalysisSettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnalysisSettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalysisSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
