% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisSettings.R
\name{AnalysisSettings}
\alias{AnalysisSettings}
\title{AnalysisSettings}
\description{
AnalysisSettings

AnalysisSettings
}
\details{
AnalysisSettings R6 object. Inherits from Settings
}
\section{Super class}{
\code{TreatmentPatterns::Settings} -> \code{AnalysisSettings}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnalysisSettings-new}{\code{AnalysisSettings$new()}}
\item \href{#method-AnalysisSettings-editSettings}{\code{AnalysisSettings$editSettings()}}
\item \href{#method-AnalysisSettings-get}{\code{AnalysisSettings$get()}}
\item \href{#method-AnalysisSettings-validate}{\code{AnalysisSettings$validate()}}
\item \href{#method-AnalysisSettings-clone}{\code{AnalysisSettings$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="TreatmentPatterns" data-topic="Settings" data-id="print"><a href='../../TreatmentPatterns/html/Settings.html#method-Settings-print'><code>TreatmentPatterns::Settings$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnalysisSettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnalysisSettings-new}{}}}
\subsection{Method \code{new()}}{
Initializer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalysisSettings$new(cohortSettings, studyName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cohortSettings}}{<CohortSettings> An instance of CohortSettings R6 object.}

\item{\code{studyName}}{<character> Name of the study.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnalysisSettings-editSettings"></a>}}
\if{latex}{\out{\hypertarget{method-AnalysisSettings-editSettings}{}}}
\subsection{Method \code{editSettings()}}{
Function to edit the analysis settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalysisSettings$editSettings(
  includeTreatments = "startDate",
  periodPriorToIndex = 0,
  minEraDuration = 0,
  splitEventCohorts = "",
  splitTime = 30,
  eraCollapseSize = 30,
  combinationWindow = 30,
  minPostCombinationDuration = 30,
  filterTreatments = "First",
  maxPathLength = 5,
  minCellCount = 5,
  minCellMethod = "Remove",
  groupCombinations = 10,
  addNoPaths = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{includeTreatments}}{<character>}

\item{\code{periodPriorToIndex}}{<numeric>}

\item{\code{minEraDuration}}{<numeric>}

\item{\code{splitEventCohorts}}{<character>}

\item{\code{splitTime}}{<numeric>}

\item{\code{eraCollapseSize}}{<numeric>}

\item{\code{combinationWindow}}{<numeric>}

\item{\code{minPostCombinationDuration}}{<numeric>}

\item{\code{filterTreatments}}{<character>}

\item{\code{maxPathLength}}{<numeric>}

\item{\code{minCellCount}}{<numeric>}

\item{\code{minCellMethod}}{<character>}

\item{\code{groupCombinations}}{<numeric>}

\item{\code{addNoPaths}}{<logical>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(self)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnalysisSettings-get"></a>}}
\if{latex}{\out{\hypertarget{method-AnalysisSettings-get}{}}}
\subsection{Method \code{get()}}{
get
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalysisSettings$get()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<list>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnalysisSettings-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AnalysisSettings-validate}{}}}
\subsection{Method \code{validate()}}{
validate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalysisSettings$validate(settings)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{settings}}{<list> settings}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(self)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnalysisSettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnalysisSettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalysisSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
