% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCohortSettings.R
\name{checkCohorts}
\alias{checkCohorts}
\title{checkCohorts}
\usage{
checkCohorts(cohorts)
}
\arguments{
\item{cohorts}{(\link[base]{data.frame})\cr
Data frame of specified cohorts.}
}
\value{
(\link[base]{logical})\cr
TRUE if checkmate checks pass.
}
\description{
Checks the validity of targetCohorts and eventCohorts parameters of
createCohortSettings
}
