% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructPathways.R
\name{checkConstructPathways}
\alias{checkConstructPathways}
\title{checkConstructPathways}
\usage{
checkConstructPathways(dataSettings, pathwaySettings, saveSettings)
}
\arguments{
\item{dataSettings}{}

\item{pathwaySettings}{}

\item{saveSettings}{}
}
\value{
TRUE if all assertions pass
}
\description{
Checks parameters for constructPathways.
}
