% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructPathways.R
\name{checkConstructPathways}
\alias{checkConstructPathways}
\title{checkConstructPathways}
\usage{
checkConstructPathways(dataSettings, pathwaySettings, saveSettings)
}
\arguments{
\item{dataSettings}{dataSettings object}

\item{pathwaySettings}{pathwaySettings object}

\item{saveSettings}{saveSettings object}
}
\value{
TRUE if all assertions pass
}
\description{
Checks parameters for constructPathways.
}
