% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDataSettings.R
\name{checkDataSettings}
\alias{checkDataSettings}
\title{checkDataSettings}
\usage{
checkDataSettings(
  connectionDetails,
  cdmDatabaseSchema,
  resultSchema,
  cohortTable
)
}
\arguments{
\item{connectionDetails}{connectionDetails object.}

\item{cdmDatabaseSchema}{Schema name of CDM.}

\item{resultSchema}{Schema name of results.}

\item{cohortTable}{Cohort table name.}
}
\value{
TRUE if all the assertions pass.
}
\description{
Check createDataSettings parameters.
}
