% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPathwaySettings.R
\name{checkPathwaySettings}
\alias{checkPathwaySettings}
\title{checkPathwaySettings}
\usage{
checkPathwaySettings(cohortSettings)
}
\arguments{
\item{cohortSettings}{cohortSettings object}
}
\value{
TRUE if all assertions pass
}
\description{
Check whether an object has the correct CohortSettings and dataframe class. The function checks it has the specified amount of columns, and if they are integer and character. Also checks if the dataframe columns have the correct name.
}
\examples{
\dontrun{
  checkPathwaySettings(
    cohortSettings = cohortSettings
    )}
}
