% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSaveSettings.R
\name{checkSaveSettings}
\alias{checkSaveSettings}
\title{checkSaveSettings}
\usage{
checkSaveSettings(env)
}
\arguments{
\item{env}{Environment containging all the function environment variables.}
}
\value{
TRUE if all assertions pass
}
\description{
Checks parameters for createSaveSettings.
}
