% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSaveSettings.R
\name{checkSaveSettings}
\alias{checkSaveSettings}
\title{checkSaveSettings}
\usage{
checkSaveSettings(databaseName, rootFolder, outputFolder, tempFolder)
}
\arguments{
\item{databaseName}{Name of the database that will appear in the results.}

\item{rootFolder}{Name of local folder to place all package output (outputFolder,
tempFolder if not given).}

\item{outputFolder}{Name of local folder to place package output.}

\item{tempFolder}{Name of local temporal folder.}
}
\value{
TRUE if all assertions pass
}
\description{
Checks parameters for createSaveSettings.
}
