% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortCharacterization.R
\name{cohortCharacterization}
\alias{cohortCharacterization}
\title{cohortCharacterization}
\usage{
cohortCharacterization(dataSettings, characterizationSettings, saveSettings)
}
\arguments{
\item{dataSettings}{Settings object as created by createDataSettings().}

\item{characterizationSettings}{Settings object as created by createCharcterizationSettings().}

\item{saveSettings}{Settings object as created by createSaveSettings().}
}
\description{
Optional, only for OMOP-CDM data: Perform baseline characterization of 
study/target population.
}
