% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Characterization.R
\name{cohortCharacterization}
\alias{cohortCharacterization}
\title{This function will perform baseline characterization for the the target cohorts.}
\usage{
cohortCharacterization(
  connection,
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  outputFolder,
  databaseId,
  targetCohortIds,
  minCellCount
)
}
\arguments{
\item{connection}{Connection to database server.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
createConnectionDetails function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data resides if OMOP-CDM = TRUE.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{cohortTable}{The name of the table that will be created in the work database schema.
This table will hold the target and event cohorts used in this
study.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/).}

\item{databaseId}{Unique identifier for database (can be the same as databaseName).}

\item{targetCohortIds}{IDs to refer to target cohorts.}

\item{minCellCount}{Minimum number of persons with a specific treatment pathway for the pathway to be included in analysis.}
}
\description{
This function will perform baseline characterization for the the target cohorts.
}
