% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{computeCounts}
\alias{computeCounts}
\title{computeCounts}
\usage{
computeCounts(treatmentHistory, minCellCount)
}
\arguments{
\item{treatmentHistory}{(\code{data.frame()})
Patient level Treatment History data.frame}

\item{minCellCount}{(\code{integer(1)}: \code{5})\cr
Minimum count required per pathway. Censors data below \code{x} as \verb{<x}. This
minimum value will carry over to the sankey diagram and sunburst plot.}
}
\value{
(\code{list()})
}
\description{
computeCounts
}
