% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{computeStatsTherapy}
\alias{computeStatsTherapy}
\title{computeStatsTherapy}
\usage{
computeStatsTherapy(treatmentHistory)
}
\arguments{
\item{treatmentHistory}{(\code{data.frame()})
Patient level Treatment History data.frame}
}
\value{
(\code{data.frame()})
}
\description{
computeStatsTherapy
}
