% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{computeTreatmentPathways}
\alias{computeTreatmentPathways}
\title{computeTreatmentPathways}
\usage{
computeTreatmentPathways(treatmentHistory, ageWindow, minFreq)
}
\arguments{
\item{treatmentHistory}{(\code{data.frame()})}

\item{ageWindow}{(\code{integer(1)})}

\item{minFreq}{(\code{integer(1)})}
}
\value{
(\code{data.frame()})
}
\description{
computeTreatmentPathways
}
