% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{computeTreatmentPathways}
\alias{computeTreatmentPathways}
\title{computeTreatmentPathways}
\usage{
computeTreatmentPathways(treatmentHistory, ageWindow, minFreq)
}
\arguments{
\item{treatmentHistory}{(\code{data.frame()})
Patient level Treatment History data.frame}

\item{ageWindow}{(\code{integer(1)}: \code{10})\cr
Number of years to bin age groups into.}

\item{minFreq}{(\code{integer(1)}: \code{5})\cr
Minimum frequency required per pathway. Censors data below \code{x} as \verb{<x}. This
minimum value will carry over to the sankey diagram and sunburst plot.}
}
\value{
(\code{data.frame()})
}
\description{
computeTreatmentPathways
}
