% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructPathways.R
\name{constructPathways}
\alias{constructPathways}
\title{This function will construct treatment pathways.}
\usage{
constructPathways(
  all_data,
  study_settings,
  databaseName,
  studyName,
  outputFolder
)
}
\arguments{
\item{all_data}{Datatable with all target and event cohorts.}

\item{study_settings}{Object that contains all study settings inputted by the user.}

\item{databaseName}{Name of the database that will appear in the results.}

\item{studyName}{Name for the study corresponding to the current settings.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes (/).}
}
\description{
This function will construct treatment pathways.
}
