% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructPathways.R
\name{constructPathways}
\alias{constructPathways}
\title{constructPathways}
\usage{
constructPathways(dataSettings, pathwaySettings, saveSettings)
}
\arguments{
\item{dataSettings}{Settings object as created by createDataSettings().}

\item{pathwaySettings}{Settings object as created by createPathwaySettings().}

\item{saveSettings}{Settings object as created by createSaveSettings().}
}
\description{
Construct treatment pathways. Also generates output in csv format.
}
\examples{
\dontrun{
  constructPathways(
    dataSettings = dataSettings,
    pathwaySettings = pathwaySettings,
    saveSettings = saveSettings)}
}
