% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructPathways.R
\name{constructPathways}
\alias{constructPathways}
\title{constructPathways}
\usage{
constructPathways(pathwaySettings, cohortSettings, andromeda)
}
\arguments{
\item{pathwaySettings}{Settings object as created by \link[TreatmentPatterns]{createPathwaySettings}.}

\item{cohortSettings}{Settings object created by \link[TreatmentPatterns]{createCohortSettings}.}

\item{dataSettings}{Settings object as created by \link[TreatmentPatterns]{createDataSettings}.}
}
\description{
Construct treatment pathways. Also generates output in csv format.
}
\examples{
\dontrun{
  constructPathways(
    dataSettings = dataSettings,
    pathwaySettings = pathwaySettings,
    saveSettings = saveSettings,
    cohortSettings = cohortSettings
  )
}
}
