% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructPathways.R
\name{constructPathways}
\alias{constructPathways}
\title{constructPathways}
\usage{
constructPathways(pathwaySettings, cohortSettings, andromeda)
}
\arguments{
\item{pathwaySettings}{(\code{TreatmentPatterns::PathwaySettings})}

\item{cohortSettings}{(\code{TreatmentPatterns::CohortSettings})}

\item{andromeda}{(\code{Andromeda::andromeda()})}
}
\value{
(\code{list}) List containing andromeda objects.
}
\description{
Runs constructPathway for each set of settings in pathwaySettings.
}
