% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructPathways.R
\name{constructPathways}
\alias{constructPathways}
\title{constructPathway}
\usage{
constructPathways(settings, cohorts, andromeda)
}
\arguments{
\item{settings}{(\code{data.frame})}

\item{cohorts}{(\code{data.frame})}

\item{andromeda}{(\code{Andromeda::andromeda()})}
}
\value{
\code{invisible(NULL)}
}
\description{
Constructs the pathways.
}
