% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSettings.R
\name{createCharacterizationSettings}
\alias{createCharacterizationSettings}
\title{Create characterization settings (optional, only for OMOP-CDM data ).}
\usage{
createCharacterizationSettings(
  baselineCovariates_location = NULL,
  baselineCovariates = data.frame(covariateName = c("Male", "Age",
    "Charlson comorbidity index score"), covariateId = c(8507001, 1002, 1901)),
 
    standardCovariateSettings = FeatureExtraction::createCovariateSettings(useDemographicsAge
    = TRUE, useDemographicsGender = TRUE, useDemographicsTimeInCohort = TRUE,
    useDemographicsPostObservationTime = TRUE, useConditionGroupEraAnyTimePrior = TRUE,
    useConditionGroupEraLongTerm = TRUE, useCharlsonIndex = TRUE),
  minCellCount = 5
)
}
\arguments{
\item{baselineCovariates_location}{Optional: Location of saved baselieCovariates object.}

\item{baselineCovariates}{Data frame containing the baseline characteristics of interest (covariateName, covariateId).}

\item{standardCovariateSettings}{An object of type covariateSettings as created using the createCovariateSettings function in the FeatureExtraction package.}
}
\value{

}
\description{
Create characterization settings (optional, only for OMOP-CDM data ).
}
