% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCohorts.R
\name{createCohorts}
\alias{createCohorts}
\title{Create target/event cohorts of interest.}
\usage{
createCohorts(dataSettings, cohortSettings, saveSettings)
}
\arguments{
\item{dataSettings}{Settings object as created by createDataSettings().}

\item{cohortSettings}{Settings object as created by createCohortSettings().}

\item{saveSettings}{Settings object as created by createSaveSettings().}
}
\description{
Create target/event cohorts of interest.
}
