% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSettings.R
\name{createCohortSettings}
\alias{createCohortSettings}
\title{Create cohort settings.}
\usage{
createCohortSettings(
  cohortsToCreate_location = NULL,
  targetCohorts = NULL,
  eventCohorts = NULL,
  loadCohorts = FALSE,
  cohortsFolder = NULL,
  baseUrl = NULL,
  generateCohorts = TRUE
)
}
\arguments{
\item{cohortsToCreate_location}{Optional: Location of saved cohortsToCreate object.}

\item{targetCohorts}{Data frame containing the study population of interest (cohortId, cohortName, atlasId, conceptSet).}

\item{eventCohorts}{Data frame containing the events of interest (cohortId, cohortName, atlasId, conceptSet).}

\item{loadCohorts}{Setting to retrieve cohort definitions with atlasId from ATLAS WebApi.}

\item{baseUrl}{The base URL for the WebApi instance, for example: "http://server.org:80/WebAPI".
Note, there is no trailing '/'. If trailing '/' is used, you may receive an error.}

\item{generateCohorts}{Setting to (re)generate cohortTable in the database.}
}
\value{

}
\description{
Create cohort settings.
}
