% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCohortSettings.R
\name{createCohortSettings}
\alias{createCohortSettings}
\title{Create cohort settings.}
\usage{
createCohortSettings(targetCohorts, eventCohorts)
}
\arguments{
\item{targetCohorts}{Data frame containing the study population of interest
cohortId = "Unique ID number", cohortName = "Descriptive name cohort".}

\item{eventCohorts}{Data frame containing the events of interest
cohortId = "Unique ID number", cohortName = "Descriptive name cohort".}
}
\value{
Object cohortSettings.
}
\description{
Create a cohortsSettings object, containing information about the target and event cohorts.
A cohort ID and name are required to specify the target and event cohorts.
The cohortId and cohortName are the ID and Name specified while creating cohorts with i.e. CohortGenerator.
}
\examples{
cohortSettings <- createCohortSettings(
  targetCohorts = data.frame(
    cohortId = c(1),
    cohortName = c("a")),
  eventCohorts = data.frame(
    cohortId = c(2, 3),
    cohortName = c("b", "c")))
}
