% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDataSettings.R
\name{createDataSettings}
\alias{createDataSettings}
\title{createDataSettings}
\usage{
createDataSettings(
  connectionDetails,
  cdmDatabaseSchema,
  resultSchema,
  cohortTable = "treatmentpatterns_cohorts"
)
}
\arguments{
\item{connectionDetails}{Only for omopCDM TRUE: An object of type connectionDetails as created
using the createConnectionDetails function in the DatabaseConnector
package.}

\item{cdmDatabaseSchema}{Only for omopCDM TRUE: Schema name where your patient-level data
resides. Note that for SQL Server, this should include both the database
and schema name, for example 'cdm_data.dbo'.}

\item{resultSchema}{Schema name of results.}

\item{cohortTable}{Only for omopCDM TRUE: The name of the table that will be created in
the cohortDatabaseSchema. This table will hold the target and event
cohorts used in this study.}
}
\value{
Object dataSettings.
}
\description{
Create a dataSettings object containing information about how to connect to
a database.
}
