% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOutput.R
\name{createLegend}
\alias{createLegend}
\title{createLegend}
\usage{
createLegend(studyName, outputFolder, databaseName)
}
\arguments{
\item{studyName}{Study name}

\item{outputFolder}{Output folder}

\item{databaseName}{Database name}
}
\description{
Write html file for the sunburst plot legend.
}
\examples{
\dontrun{
source(system.file(
  package = "TreatmentPatterns",
  "testing", "testParamsOutput.R"))

TreatmentPatterns:::createLegend(
  studyName = "Viral_Sinusitis",
  outputFolder = saveSettings$outputFolder,
  databaseName = saveSettings$databaseName)
}
}
