% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSettings.R
\name{createPathwaySettings}
\alias{createPathwaySettings}
\title{Create pathway settings.}
\usage{
createPathwaySettings(
  pathwaySettings_location = NULL,
  pathwaySettings_list = NULL,
  targetCohortId = NULL,
  eventCohortIds = NULL
)
}
\arguments{
\item{pathwaySettings_location}{Optional: Location of saved pathwaySettings object.}

\item{pathwaySettings_list}{Create (list of pathway settings) with addPathwaySettings()
(e.g.pathwaySettings_list <- addPathwaySettings() or pathwaySettings_list <- list(addPathwaySettings(), addPathwaySettings())). .}

\item{targetCohortId}{Target cohort ID of current study settings.}

\item{eventCohortIds}{Event cohort IDs of current study settings.}
}
\value{
Object pathwaySettings.
}
\description{
Create pathway settings.
}
