% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPathwaySettings.R
\name{createPathwaySettings}
\alias{createPathwaySettings}
\title{createPathwaySettings}
\usage{
createPathwaySettings(cohortSettings, ...)
}
\arguments{
\item{cohortSettings}{cohortSettings object}

\item{...}{Any addPathwaySettings parameter:
1. studyName
2. includeTreatments
3. periodPriorToIndex
4. minEraDuration
5. splitEventCohorts
6. splitTime
7. eraCollapseSize
8. combinationWindow
9. minPostCombinationDuration
10. filterTreatments
11. maxPathLength
12. minCellCount
13. minCellMethod
14. groupCombinations
15. addNoPaths}
}
\value{
Object pathwaySettings.
}
\description{
Create pathway settings.
}
\examples{
targetCohorts <- data.frame(
  cohortName = c("targetCohort"),
  cohortId = c(1))

eventCohorts <- data.frame(
  cohortName = c("eventCohort1", "eventCohort2"),
  cohortId = c(2, 3))

cohortSettings <- TreatmentPatterns::createCohortSettings(
  targetCohorts = targetCohorts,
  eventCohorts = eventCohorts)

createPathwaySettings(
  cohortSettings = cohortSettings,
  studyName = "MyStudyName")
}
