% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPathwaySettings.R
\name{createPathwaySettings}
\alias{createPathwaySettings}
\title{createPathwaySettings}
\usage{
createPathwaySettings(cohortSettings, ...)
}
\arguments{
\item{cohortSettings}{(\link[TreatmentPatterns]{addPathwaySettings})\cr
cohortSettings object}

\item{...}{\enumerate{
\item studyName
\item includeTreatments
\item periodPriorToIndex
\item minEraDuration
\item splitEventCohorts
\item splitTime
\item eraCollapseSize
\item combinationWindow
\item minPostCombinationDuration
\item filterTreatments
\item maxPathLength
\item minCellCount
\item minCellMethod
\item groupCombinations
\item addNoPaths
}}
}
\value{
(\link[TreatmentPatterns]{createPathwaySettings})
S3 pathwaySettings object.
}
\description{
Create pathway settings.
}
\examples{
targetCohorts <- data.frame(
  cohortName = c("targetCohort"),
  cohortId = c(1))

eventCohorts <- data.frame(
  cohortName = c("eventCohort1", "eventCohort2"),
  cohortId = c(2, 3))

cohortSettings <- TreatmentPatterns::createCohortSettings(
  targetCohorts = targetCohorts,
  eventCohorts = eventCohorts)

createPathwaySettings(
  cohortSettings = cohortSettings,
  studyName = "MyStudyName")
}
