% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSankeyDiagram.R
\name{createSankeyDiagram}
\alias{createSankeyDiagram}
\title{createSankeyDiagram}
\usage{
createSankeyDiagram(
  data,
  outputFolder,
  groupCombinations,
  fileName = "sankeyDiagram.html"
)
}
\arguments{
\item{data}{Dataframe with event cohorts of the target cohort in different columns.}

\item{outputFolder}{Path to the output folder.}

\item{groupCombinations}{Select to group all non-fixed combinations in one category "other" in
the sunburst plot.}

\item{fileName}{File name of the html-file to output.}
}
\description{
createSankeyDiagram
}
