% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSankeyDiagram.R
\name{createSankeyDiagram}
\alias{createSankeyDiagram}
\title{createSankeyDiagram}
\usage{
createSankeyDiagram(
  treatmentHistory,
  outputFile,
  groupCombinations = FALSE,
  minFreq = 5
)
}
\arguments{
\item{treatmentHistory}{(\code{data.frame()})\cr
Treatment history data frame.}

\item{outputFile}{(\code{character(1)})\cr
Path where the Sankey diagram should be written to.}

\item{groupCombinations}{(\code{logical(1)}: \code{TRUE})\cr
\describe{
\item{\code{TRUE}}{Group all combination treatments in category \code{"Combination"}.}
\item{\code{FALSE}}{Do not group combination treatments.}
}}

\item{minFreq}{(\code{integer(1)}: \code{5})\cr
Censor paths with a frequency smaller than specified.}
}
\value{
invisible(NULL)
}
\description{
Writes the Sankey diagram to a HTML-file, to a specified file path.
}
\examples{
if (FALSE) {
  TH <- andromeda$myStudy$treatmentHistory \%>\% collect()

  createSankeyDiagram(
    treatmentHistory = TH,
    outputFile = "./some/path/to/sankeyDiagram.html",
    groupCombinations = FALSE,
    minFreq = 0
  )
}
}
