% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSankeyDiagram.R
\name{createSankeyDiagram}
\alias{createSankeyDiagram}
\title{createSankeyDiagram}
\usage{
createSankeyDiagram(
  data,
  studyName = NULL,
  outputFolder,
  groupCombinations,
  fileName = "sankeyDiagram.html"
)
}
\arguments{
\item{data}{(\link[base]{data.frame})
Data frame with event cohorts of the target cohort in different columns.}

\item{studyName}{(\link[base]{character})
Name of the study, specified in the \code{pathwaySerttings} object.}

\item{outputFolder}{(\link[base]{character})
Path to the output folder.}

\item{groupCombinations}{(\link[base]{logical})
Select to group all non-fixed combinations in one category "other" in the sunburst plot.}

\item{fileName}{(\link[base]{character})
File name of the html-file to output.}
}
\description{
createSankeyDiagram
}
\examples{
\dontrun{
  creaeSankeyDiagram(
    data = read.csv("freqPerYear.csv"),
    studyName = "Viral_Sinusitis",
    groupCombinations = FALSE,
    outputFolder = "output")
}
}
