% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSankeyDiagram.R
\name{createSankeyDiagram}
\alias{createSankeyDiagram}
\title{createSankeyDiagram}
\usage{
createSankeyDiagram(
  treatmentPathways,
  outputFile,
  groupCombinations = FALSE,
  minFreq = 5,
  year = "all"
)
}
\arguments{
\item{treatmentPathways}{(\code{data.frame()})\cr
Data frame containing treatmentPathways columsn: path, freq.}

\item{outputFile}{(\code{character(1)})\cr
Path where the Sankey diagram should be written to.}

\item{groupCombinations}{(\code{logical(1)}: \code{FALSE})\cr
\describe{
\item{\code{TRUE}}{Group all combination treatments in category \code{"Combination"}.}
\item{\code{FALSE}}{Do not group combination treatments.}
}}

\item{minFreq}{(\code{integer(1)}: \code{5})\cr
Minimum frequency required per pathway. Censors data below \code{x} as \verb{<x}. This
minimum value will carry over to the sankey diagram and sunburst plot.}

\item{year}{(\code{integer(1)} | \code{character(1)})\cr
Year}
}
\value{
invisible(NULL)
}
\description{
Writes the Sankey diagram to a HTML-file, to a specified file path.
}
