% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSaveSettings.R
\name{createSaveSettings}
\alias{createSaveSettings}
\title{createSaveSettings}
\usage{
createSaveSettings(
  databaseName = "unknown_name",
  rootFolder,
  outputFolder = file.path(rootFolder, "output"),
  tempFolder = file.path(rootFolder, "temp")
)
}
\arguments{
\item{databaseName}{Name of the database that will appear in the results.}

\item{rootFolder}{Name of local folder to place all package output (outputFolder,
tempFolder if not given).}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/).}

\item{tempFolder}{Name of local folder to place intermediate results (not to be shared);
make sure to use forward slashes (/).}
}
\value{
Object saveSettings.
}
\description{
Creates saveSettings object.
}
