% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSaveSettings.R
\name{createSaveSettings}
\alias{createSaveSettings}
\title{createSaveSettings}
\usage{
createSaveSettings(
  databaseName = "unknown_name",
  rootFolder,
  outputFolder = NULL,
  tempFolder = NULL
)
}
\arguments{
\item{databaseName}{Name of the database that will appear in the results.}

\item{rootFolder}{Name of local folder to place all package output (outputFolder,
tempFolder if not given).}
}
\value{
Object saveSettings.
}
\description{
Creates saveSettings object.
}
