% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSaveSettings.R
\name{createSaveSettings}
\alias{createSaveSettings}
\title{createSaveSettings}
\usage{
createSaveSettings(
  databaseName = "unknown_name",
  rootFolder,
  outputFolder = NULL,
  tempFolder = NULL
)
}
\arguments{
\item{databaseName}{Name of the database that will appear in the results.}

\item{rootFolder}{Name of local folder to place all package output (outputFolder,
tempFolder if not given).}

\item{outputFolder}{Name of local folder to place package output.}

\item{tempFolder}{Name of local temporal folder.}
}
\value{
Object saveSettings.
}
\description{
Creates saveSettings object.
}
\examples{
createSaveSettings(
   rootFolder = "./",
   databaseName = "Eunomia",
   outputFolder = "./inst",
   tempFolder = "./inst")
}
