% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOutput.R
\name{createSunburstPlot}
\alias{createSunburstPlot}
\title{createSunburstPlot}
\usage{
createSunburstPlot(
  data,
  outcomes = NULL,
  folder = NULL,
  fileName = NULL,
  shiny = FALSE,
  title = ""
)
}
\arguments{
\item{data}{A data frame containing two columns: 1) column "path" should specify the
event cohorts separated by dashes - (combinations can be indicated using
&) and 2) column "freq" should specify how often that (unique) path
occurs.}

\item{outcomes}{Character vector containing all event cohorts.}

\item{folder}{Root folder to store the results.}

\item{fileName}{File name for the results.}

\item{shiny}{Set to TRUE if HTML file is generated for shiny application, FALSE will
generate a standalone HTML with title and legend.}

\item{title}{Optional if shiny = FALSE: add descriptive title in sunburst plot for
standalone HTML.}
}
\description{
Export a sunburst plot from a data.frame object.
}
\examples{
\dontrun{
createSunburstPlot(
  data = data.frame(
    path = c("1", "2"),
    freq = c("0.5", "0.5")))
}
}
