% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSunburstPlot.R
\name{createSunburstPlot}
\alias{createSunburstPlot}
\title{createSunburstPlot}
\usage{
createSunburstPlot(data, folder, fileName)
}
\arguments{
\item{data}{A data frame containing two columns: 1) column "path" should specify the
event cohorts separated by dashes - (combinations can be indicated using
&) and 2) column "freq" should specify how often that (unique) path
occurs.}

\item{folder}{Root folder to store the results.}

\item{fileName}{File name for the results.}
}
\description{
Export a sunburst plot from a data.frame object.
}
\examples{
\dontrun{
createSunburstPlot(
  data = data.frame(
    path = c("1", "2"),
    freq = c("0.5", "0.5")))
}
}
