% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSunburstPlot.R
\name{createSunburstPlot}
\alias{createSunburstPlot}
\title{createSunburstPlot}
\usage{
createSunburstPlot(treatmentHistory, outputFile, year = "all")
}
\arguments{
\item{treatmentHistory}{(\code{data.frame()})\cr
Treatment history data.frame from an \code{Andromeda::andromeda()} environment,
obtained by calling \code{pathwayConstructor$getAndromedas()}.}

\item{outputFile}{(\code{character(1)})\cr
Path to output file.}

\item{year}{(\code{integer(1)})\cr
Year to filter on. May also be \code{"all"}.}
}
\description{
Export a sunburst plot from a data.frame object.
}
