% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSunburstPlot.R
\name{createSunburstPlot}
\alias{createSunburstPlot}
\title{createSunburstPlot}
\usage{
createSunburstPlot(treatmentPathways, outputFile)
}
\arguments{
\item{treatmentPathways}{(\code{data.frame()})\cr
The contents of the treatmentPathways.csv-file as a data.frame().}

\item{outputFile}{(\code{character(1)})\cr
Path where the Sankey diagram should be written to.}
}
\value{
(\code{NULL})
}
\description{
Generate a sunburst plot from the treatment pathways.
}
\examples{
if (FALSE) {
  treatmentPathways <- read.csv("treatmentPathways.csv")
  outputFile <- "mySunburstPlot.html"

  createSunburstPlot(
    treatmentPatwhays,
    outputFile
  )
}
}
