% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSunburstPlot.R
\name{createSunburstPlot}
\alias{createSunburstPlot}
\title{createSunburstPlot}
\usage{
createSunburstPlot(treatmentPathways, outputFile)
}
\arguments{
\item{treatmentPathways}{(\code{data.frame()})\cr
Data frame containing treatmentPathways columns: path, freq.}

\item{outputFile}{(\code{character(1)})\cr
Path to output file.}
}
\value{
(\code{NULL})
}
\description{
Export a sunburst plot from a data.frame object.
}
