% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructPathways.R
\name{createTreatmentHistory}
\alias{createTreatmentHistory}
\title{createTreatmentHistory}
\usage{
createTreatmentHistory(
  andromeda,
  targetCohortIds,
  eventCohortIds,
  exitCohortIds,
  periodPriorToIndex,
  includeTreatments
)
}
\arguments{
\item{andromeda}{(\code{Andromeda::andromeda()})
Andromeda object.}

\item{targetCohortIds}{(\code{numeric(n)})}

\item{eventCohortIds}{(\code{numeric(n)})}

\item{exitCohortIds}{(\code{numeric(n)})}

\item{periodPriorToIndex}{(\code{integer(1)}: \code{0})\cr
Number of days prior to the index date of the target cohort | that event
cohorts are allowed to start}

\item{includeTreatments}{(\code{character(1)}: \code{"startDate"})\cr}
}
\value{
(\code{data.frame()})\cr
\enumerate{
\item (\code{numeric()}) personId
\item (\code{numeric()}) indexYear
\item (\code{numeric()}) eventCohortId
\item (\code{as.Date()}) eventStartDate
\item (\code{as.Date()}) eventEndDate
\item (\code{character()}) type
\item (\code{difftime()}) durationEra
\item (\code{difftime()}) gapSame
}
}
\description{
createTreatmentHistory
}
