% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darwinExecute.R
\name{darwinExecute}
\alias{darwinExecute}
\title{darwinExecute}
\usage{
darwinExecute(dataSettings, pathwaySettings, saveSettings)
}
\arguments{
\item{dataSettings}{dataSettings object}

\item{pathwaySettings}{pathwaySettings object}

\item{saveSettings}{saveSettings object}
}
\description{
darwinExecute
}
\examples{
if (interactive()) {
  # Select Viral Sinusitis Cohort
  targetCohort <- cohortsGenerated \%>\% 
    filter(cohortName == "Viral Sinusitis") \%>\%
    select(cohortId, cohortName)

  # Select everything BUT Viral Sinusitis cohorts
  eventCohorts <- cohortsGenerated \%>\% 
    filter(cohortName != "Viral Sinusitis") \%>\%
    select(cohortId, cohortName)

  saveSettings <- TreatmentPatterns::createSaveSettings(
    databaseName = "Eunomia",
    rootFolder = getwd(),
    outputFolder = file.path(getwd(), "output", "Eunomia"))
  
  cohortSettings <- TreatmentPatterns::createCohortSettings(
    targetCohorts = targetCohort,
     eventCohorts = eventCohorts)

  pathwaySettings <- createPathwaySettings(
    cohortSettings = cohortSettings,
    studyName = "Viral_Sinusitis")
  
  darwinExecute(
  dataSettings = dataSettings,
  pathwaySettings = pathwaySettings,
  saveSettings = saveSettings)
}
}
