% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-DataSettings.R
\name{DataSettings}
\alias{DataSettings}
\title{R6 DataSettings class}
\description{
DataSettings object. Inherits from Settings
}
\examples{
connectionDetails <- Eunomia::getEunomiaConnectionDetails()

DataSettings$new(
  connectionDetails = connectionDetails,
  cdmDatabaseSchema = "main",
  resultSchema = "main",
  cohortTable = "cohortTable"
)
}
\seealso{
Other Settings: 
\code{\link{AnalysisSettings}},
\code{\link{CohortSettings}},
\code{\link{PathwaySettings}},
\code{\link{SaveSettings}}
}
\concept{Settings}
\section{Super class}{
\code{TreatmentPatterns::Settings} -> \code{DataSettings}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataSettings-new}{\code{DataSettings$new()}}
\item \href{#method-DataSettings-print}{\code{DataSettings$print()}}
\item \href{#method-DataSettings-validate}{\code{DataSettings$validate()}}
\item \href{#method-DataSettings-get}{\code{DataSettings$get()}}
\item \href{#method-DataSettings-clone}{\code{DataSettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataSettings-new}{}}}
\subsection{Method \code{new()}}{
[Overload] Initializer method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSettings$new(
  connectionDetails,
  cdmDatabaseSchema,
  resultSchema,
  cohortTable
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connectionDetails}}{(\link[DatabaseConnector]{createConnectionDetails})
ConnectionDetails object.}

\item{\code{cdmDatabaseSchema}}{(\link[base]{character})
CDM Schema in database.}

\item{\code{resultSchema}}{(\link[base]{character})
Result Schema in database.}

\item{\code{cohortTable}}{(\link[base]{character})
Name of the cohort table.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSettings-print"></a>}}
\if{latex}{\out{\hypertarget{method-DataSettings-print}{}}}
\subsection{Method \code{print()}}{
Overload print
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSettings$print(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{DataSettings object.}

\item{\code{...}}{Other parameters for \link[base]{print}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Message from \link[base]{cat}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSettings-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DataSettings-validate}{}}}
\subsection{Method \code{validate()}}{
!overload validation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSettings$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSettings-get"></a>}}
\if{latex}{\out{\hypertarget{method-DataSettings-get}{}}}
\subsection{Method \code{get()}}{
Gets the specified items packged in a named list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSettings$get()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\link[base]{list})
\describe{
\item{connectionDetails}{(\code{ConnectionDetails}) ConnectionDetails object. See \link[DatabaseConnector]{createConnectionDetails}.}
\item{cdmDatabaseSchema}{(\link[base]{character}) Name of the CDM Schema in database.}
\item{resultSchema}{(\link[base]{character}) Name of the Result Schema in the database.}
\item{cohortTable}{(\link[base]{character}) Name of the cohort table}
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataSettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
