% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructPathways.R
\name{doCombinationWindow}
\alias{doCombinationWindow}
\title{Combine overlapping events into combinations}
\usage{
doCombinationWindow(andromeda, combinationWindow, minPostCombinationDuration)
}
\arguments{
\item{andromeda}{(\code{Andromeda::andromeda()})}

\item{combinationWindow}{(\code{integer(1)})}

\item{minPostCombinationDuration}{(\code{integer(1)})}
}
\value{
(\code{invisible(NULL)})
}
\description{
doCombinationWindow is an internal function that combines overlapping events
into combination events. It accepts a treatmentHistory dataframe and returns
a modified treatmentHistory dataframe. The returned treatmentHistory
dataframe always has the property that a person is only in one event cohort,
which might be a combination event cohort, at any point time.
}
