% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructPathways.R
\name{doCreateTreatmentHistory}
\alias{doCreateTreatmentHistory}
\title{doCreateTreatmentHistory}
\usage{
doCreateTreatmentHistory(
  currentCohorts,
  targetCohortId,
  eventCohortIds,
  exitCohortIds,
  periodPriorToIndex,
  includeTreatments
)
}
\arguments{
\item{currentCohorts}{(\link[base]{data.frame})\cr
\enumerate{
\item (\link[base]{numeric}) cohort_id
\item (\link[base]{numeric}) person_id
\item (\link[base]{date}: \verb{\\\%Y-\\\%m-\\\%d}) start_date
\item (\link[base]{date}: \verb{\\\%Y-\\\%m-\\\%d}) end_date
}}

\item{targetCohortId}{(\link[base]{c}) of (\link[base]{numeric})\cr
targetCohortId from \link[TreatmentPatterns]{addPathwaySettings}.}

\item{eventCohortIds}{(\link[base]{c}) of (\link[base]{numeric})\cr
eventCohortIds from \link[TreatmentPatterns]{addPathwaySettings}.}

\item{exitCohortIds}{(\link[base]{c}) of (\link[base]{numeric})\cr
exitCohortIds from \link[TreatmentPatterns]{addPathwaySettings}.}

\item{periodPriorToIndex}{(\link[base]{numeric})\cr
periodPriorToIndex from \link[TreatmentPatterns]{addPathwaySettings}.}

\item{includeTreatments}{(\link[base]{character})\cr
includeTreatments from \link[TreatmentPatterns]{addPathwaySettings}.}
}
\value{
(\link[base]{data.frame})\cr
\enumerate{
\item (\link[base]{numeric}) person_id
\item (\link[base]{numeric}) index_year
\item (\link[base]{numeric}) event_cohort_id
\item (\link[base]{date}) event_start_date
\item (\link[base]{date}) event_end_date
\item (\link[base]{character}) type
\item (\link[base]{difftime}) duration_era
\item (\link[base]{difftime}) gap_same
}
}
\description{
doCreateTreatmentHistory
}
