% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructPathways.R
\name{doEraCollapse}
\alias{doEraCollapse}
\title{doEraCollapse}
\usage{
doEraCollapse(treatment_history, eraCollapseSize)
}
\arguments{
\item{treatment_history}{Dataframe with event cohorts of the target cohort in different rows.}

\item{eraCollapseSize}{Window of time between which two eras of the same event cohort are
collapsed into one era.}
}
\value{
treatment_history
    Updated dataframe, where event cohorts with
    gap_same < eraCollapseSize are collapsed.
}
\description{
doEraCollapse
}
