% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructPathways.R
\name{doEraCollapse}
\alias{doEraCollapse}
\title{doEraCollapse}
\usage{
doEraCollapse(treatment_history, eraCollapseSize)
}
\arguments{
\item{treatment_history}{Dataframe with event cohorts of the target cohort in different rows.}

\item{eraCollapseSize}{Window of time between which two eras of the same event cohort are
collapsed into one era.}
}
\value{
treatment_history
    Updated dataframe, where event cohorts with
    gap_same < eraCollapseSize are collapsed.
}
\description{
Updates the treatmentHistory data.frame where if gapSame is smaller than the specified era collapse size (eraCollapseSize) are collapsed
}
\examples{
\dontrun{
th <- doCreateTreatmentHistory(current_cohorts = currentCohorts,
                               targetCohortId = targetCohortId,
                               eventCohortIds = eventCohortIds,
                               periodPriorToIndex = periodPriorToIndex,
                               includeTreatments = includeTreatments)
doEraCollapse(treatment_history = th, eraCollapseSize = 1)
}
}
