% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructPathways.R
\name{doEraDuration}
\alias{doEraDuration}
\title{doEraDuration}
\usage{
doEraDuration(treatmentHistory, minEraDuration)
}
\arguments{
\item{treatmentHistory}{Dataframe with event cohorts of the target cohort in different rows.}

\item{minEraDuration}{Minimum time an event era should last to be included in analysis.}
}
\value{
treatmentHistory
    Updated dataframe, rows with duration <
    minEraDuration filtered out.
}
\description{
Filters the treatmentHistory based on the specified minimum era duration
}
\examples{
\dontrun{
th <- doCreateTreatmentHistory(current_cohorts = currentCohorts,
                               targetCohortId = targetCohortId,
                               eventCohortIds = eventCohortIds,
                               periodPriorToIndex = periodPriorToIndex,
                               includeTreatments = includeTreatments)
doEraDuration(treatmentHistory = th, minEraDuration = 1)
}
}
