% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructPathways.R
\name{doEraDuration}
\alias{doEraDuration}
\title{doEraDuration}
\usage{
doEraDuration(treatmentHistory, minEraDuration)
}
\arguments{
\item{treatmentHistory}{(\link[base]{data.frame})\cr
See \link[TreatmentPatterns]{doCreateTreatmentHistory}.}

\item{minEraDuration}{(\link[base]{numeric})\cr
minEraDuration from \link[TreatmentPatterns]{addPathwaySettings}.}
}
\value{
(\link[base]{data.frame})
Updated treatmentHistory dataframe, rows with duration < minEraDuration
filtered out.
}
\description{
Filters the treatmentHistory based on the specified minimum era duration
}
\examples{
\dontrun{
th <- doCreateTreatmentHistory(current_cohorts = currentCohorts,
                               targetCohortId = targetCohortId,
                               eventCohortIds = eventCohortIds,
                               periodPriorToIndex = periodPriorToIndex,
                               includeTreatments = includeTreatments)
doEraDuration(treatmentHistory = th, minEraDuration = 1)
}
}
