% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructPathways.R
\name{doEraDuration}
\alias{doEraDuration}
\title{doEraDuration}
\usage{
doEraDuration(andromeda, minEraDuration)
}
\arguments{
\item{andromeda}{(\code{Andromeda::andromeda()})}

\item{minEraDuration}{(\code{integer(1)})}
}
\value{
(\code{invisible(NULL)})
}
\description{
doEraDuration
}
