% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructPathways.R
\name{doFilterTreatments}
\alias{doFilterTreatments}
\title{doFilterTreatments}
\usage{
doFilterTreatments(treatment_history, filterTreatments)
}
\arguments{
\item{treatment_history}{Dataframe with event cohorts of the target cohort in different rows.}

\item{filterTreatments}{Select first occurrence of ('First') / changes between ('Changes') / all
event cohorts ('All').}
}
\value{
treatment_history
    Updated dataframe, where the desired event cohorts are maintained for
    the visualizations.
}
\description{
Updates the treatmentHistory data.frame where the desired event cohorts are maintained for the visualizations
}
\examples{
\dontrun{
th <- doCreateTreatmentHistory(current_cohorts = currentCohorts,
                               targetCohortId = targetCohortId,
                               eventCohortIds = eventCohortIds,
                               periodPriorToIndex = periodPriorToIndex,
                               includeTreatments = includeTreatments)

doFilterTreatments(treatment_history = th, filterTreatments = "All")}
}
