% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructPathways.R
\name{doFilterTreatments}
\alias{doFilterTreatments}
\title{doFilterTreatments}
\usage{
doFilterTreatments(andromeda, filterTreatments)
}
\arguments{
\item{andromeda}{(\code{Andromeda::andromeda()})}

\item{filterTreatments}{(\code{character(1)})}
}
\value{
(\code{invisible(NULL)})
}
\description{
Updates the treatmentHistory data.frame where the desired event cohorts are
maintained for the visualizations
}
