% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructPathways.R
\name{doMaxPathLength}
\alias{doMaxPathLength}
\title{doMaxPathLength}
\usage{
doMaxPathLength(treatment_history, maxPathLength)
}
\arguments{
\item{treatment_history}{Dataframe with event cohorts of the target cohort in different rows.}

\item{maxPathLength}{Maximum number of steps included in treatment pathway.}
}
\value{
treatment_history
Updated dataframe, where the desired event cohorts all have a seq value of <= 
maxPathLength
}
\description{
Filters the treatmentHistory data.frame where eventSeq is smaller or equal than maxPathLength
}
\examples{
\dontrun{
th <- doCreateTreatmentHistory(current_cohorts = currentCohorts,
                               targetCohortId = targetCohortId,
                               eventCohortIds = eventCohortIds,
                               periodPriorToIndex = periodPriorToIndex,
                               includeTreatments = includeTreatments)
doMaxPathLength(treatment_history = th, maxPathLength = 1)}
}
