% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructPathways.R
\name{doMaxPathLength}
\alias{doMaxPathLength}
\title{doMaxPathLength}
\usage{
doMaxPathLength(andromeda, maxPathLength)
}
\arguments{
\item{maxPathLength}{Maximum number of steps included in treatment pathway.}

\item{treatmentHistory}{Dataframe with event cohorts of the target cohort in different rows.}
}
\value{
treatmentHistory
Updated dataframe, where the desired event cohorts all have a seq value of <=
maxPathLength
}
\description{
\code{DEPRECATED}\cr
Filters the treatmentHistory data.frame where eventSeq is smaller or equal
than maxPathLength
}
