% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructPathways.R
\name{doSplitEventCohorts}
\alias{doSplitEventCohorts}
\title{doSplitEventCohorts}
\usage{
doSplitEventCohorts(andromeda, splitEventCohorts, splitTime)
}
\arguments{
\item{andromeda}{(\code{Andromeda::andromeda()})}

\item{splitEventCohorts}{(\code{character(n)})}

\item{splitTime}{(\code{integer(1)})}
}
\value{
(\code{invisible(NULL)})
}
\description{
Splits the treatmentHistory data.frame based on event cohorts into ‘acute’
and ‘therapy’ cohorts.
}
