% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeTreatmentPatterns.R
\name{executeTreatmentPatterns}
\alias{executeTreatmentPatterns}
\title{executeTreatmentPatterns}
\usage{
executeTreatmentPatterns(dataSettings, pathwaySettings, saveSettings)
}
\arguments{
\item{dataSettings}{dataSettings object}

\item{pathwaySettings}{pathwaySettings object}

\item{saveSettings}{saveSettings object}
}
\description{
This is the main function which runs all parts of the treatment pathways
analysis. The following tasks are performed sequentially: 1) Construct
treatment pathways, 2) Generate output (sunburst plots, Sankey diagrams and
more), 3) Launch shiny application to visualize the results.
}
\examples{
if (interactive()) {
  # Select Viral Sinusitis Cohort
  targetCohort <- cohortsGenerated \%>\% 
    filter(cohortName == "Viral Sinusitis") \%>\%
    select(cohortId, cohortName)

  # Select everything BUT Viral Sinusitis cohorts
  eventCohorts <- cohortsGenerated \%>\% 
    filter(cohortName != "Viral Sinusitis") \%>\%
    select(cohortId, cohortName)

  saveSettings <- TreatmentPatterns::createSaveSettings(
    databaseName = "Eunomia",
    rootFolder = getwd(),
    outputFolder = file.path(getwd(), "output", "Eunomia"))
  
  cohortSettings <- TreatmentPatterns::createCohortSettings(
    targetCohorts = targetCohort,
     eventCohorts = eventCohorts)

  pathwaySettings <- createPathwaySettings(
    cohortSettings = cohortSettings,
    studyName = "Viral_Sinusitis")
  
  executeTreatmentPatterns(
  dataSettings = dataSettings,
  pathwaySettings = pathwaySettings,
  saveSettings = saveSettings)
}
}
