% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{extractFile}
\alias{extractFile}
\title{extractFile}
\usage{
extractFile(connection, tableName, resultsSchema)
}
\arguments{
\item{connection}{\link[DatabaseConnector]{connect}\cr
Connection object from DatabaseConnector.}

\item{tableName}{\link[base]{character}\cr
Name of table.}

\item{resultsSchema}{\link[base]{character}\cr
Schema of results.
\link[DatabaseConnector]{createConnectionDetails}.}
}
\value{
\link[base]{data.frame}
The extracted table as a data frame.
}
\description{
Extract table with specific name from database server.
}
\examples{
\dontrun{
  con <- DatabaseConnector::connect(Eunomia::getEunomiaConnectionDetails())
  extractFile(con, "person", "main", "sqlite")
}
}
