% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchCohortTable.R
\name{fetchCohortTable}
\alias{fetchCohortTable}
\title{fetchCohortTable}
\usage{
fetchCohortTable(
  cdm = NULL,
  connectionDetails = NULL,
  connection = NULL,
  cdmSchema = NULL,
  writeSchema = NULL,
  cohorts,
  cohortTables
)
}
\arguments{
\item{cdm}{(\code{cdm_reference}) A CDM reference object.}

\item{connectionDetails}{(\code{ConnectionDetails}) Connection details to
establish a database connection with using DatabaseConnector.}

\item{connection}{(\code{connection}) An already established connection to the databas.}

\item{cdmSchema}{(\code{character(1)}) Schema where the OMOP CDM resides.}

\item{writeSchema}{(\code{character(1)}) Schema where to write temporary tables to, and where the \code{cohortTable} exists.}

\item{cohorts}{(\code{data.frame}) A data.farme containing atleast the columns
\code{cohortId}, \code{cohortName}, and \code{type}. \code{cohortId}
refers to the \code{cohort_definition_id} in the cohort table in the database.
\code{cohortName} refers to the name of the cohort. These two columns are
usually parsed out from a 'cohort set'. The \code{type} column refers to what
'type' a cohort is. Either \code{"event"} or \code{"target"}. An \code{"event"} cohort
will be used to build the pathways with. The \code{"target"} cohort can be
compared to an \code{indication} cohort (like in CohortMethod or
SelfControlledCaseSeries). This cohort is used to nest the events, to make
sure the events only happen during a period of interrest (i.e. some disease
diagnosis).}

\item{cohortTables}{(\code{character(n)}) Character vector of the names of the
cohort tables to use. May be more than one table name.}
}
\value{
\code{Andromeda}
}
\description{
fetchCohortTable
}
\examples{
if (interactive()) {
  cohorts <- data.frame(
    cohort_definition_id = c(1, 2, 3),
    cohort_name = c("A", "B", "C"),
    type = c("event", "event", "target"),
  )

  fetchCohortTable(
    cdm = cdm,
    cohorts = cohorts,
    cohortTables = c(cohort_table_1, cohort_table_2),
  )

  fetchCohortTable(
    cdm = cdm,
    cohorts = cohorts,
    cohortTables = c(cohort_table_1, cohort_table_2),
  )
}
}
