% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatSunburstData.R
\name{formatSunburstData}
\alias{formatSunburstData}
\title{formatSunburstData}
\usage{
formatSunburstData(filePath, indexYear, studyName = NULL)
}
\arguments{
\item{filePath}{(\link[base]{character})
Path to file containing \code{event_cohort_name1}, \code{event_cohort_name2},
\code{event_cohort_name3}, \code{event_cohort_name4}, \code{event_cohort_name5} columns.}

\item{indexYear}{(\link[base]{integer} | \link[base]{character})
Index year to filter on. "no year" may be specified, or the year as an integer.}

\item{studyName}{(\link[base]{character})
Name of the study, specified in the \code{pathwaySerttings} object.}
}
\value{
\link[base]{data.frame}
| column | data type              |
| path   | \link[base]{character} |
| freq   | \link[base]{numeric}   |
}
\description{
Function to format data from files containing event_cohort_name columns.
}
\examples{
\dontrun{
  pathData <- formatSunburstData(
    filePath = "output/treatmentPathways.csv",
    indexYear = NA,
    studyName = "Viral_Sinusitis"
  )
  
  createSunburstPlot(
    data = pathData,
    folder = "output/",
    fileName = "sunburst.html"
  )
}
}
