% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOutput.R
\name{generateOutput}
\alias{generateOutput}
\title{Generate output (sunburst plots, Sankey diagrams and more).}
\usage{
generateOutput(saveSettings)
}
\arguments{
\item{saveSettings}{Settings object as created by createSaveSettings().}
}
\description{
Generate output (sunburst plots, Sankey diagrams and more).
}
