% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOutput.R
\name{generateOutput}
\alias{generateOutput}
\title{generateOutput}
\usage{
generateOutput(saveSettings)
}
\arguments{
\item{saveSettings}{Settings object as created by createSaveSettings().}
}
\value{
some stuff
}
\description{
generateOutput
}
