% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputPathways.R
\name{generateResults}
\alias{generateResults}
\title{This function will generate all result files and plots.}
\usage{
generateResults(study_settings, databaseName, studyName, outputFolder)
}
\arguments{
\item{study_settings}{Object that contains all study settings inputted by the user.}

\item{databaseName}{Name of the database that will appear in the results.}

\item{studyName}{Name for the study corresponding to the current settings.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes (/).}
}
\description{
This function will generate all result files and plots.
}
