% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggSunburst.R
\name{ggSunburst}
\alias{ggSunburst}
\title{ggSunburst}
\usage{
ggSunburst(
  treatmentPathways,
  minFreq = 0,
  strataX = "",
  strataY = "",
  style = "default"
)
}
\arguments{
\item{treatmentPathways}{(\code{data.frame}) Treatment Pathways result containing
at least a 'pathway' and 'freq' column.}

\item{minFreq}{(\code{numeric(1)}) A minimum frequency of pathways to plot.}

\item{strataX}{(\code{character(n)}) Column names to facet by horizontally.}

\item{strataY}{(\code{character(n)}) Column names to facet by vertically.}

\item{style}{(\code{character(1)}) Name of the style to use. Right now only
\code{"darwin"} is supported. Anything else will default to the default styling.
Themes can be changed by overwriting it with any of the ggplot2 theme
functions (i.e. \code{ggSunburst(df) + ggplot2::theme_classic()}).}
}
\value{
\code{ggplot}
}
\description{
ggplot2 implementation of the sunburst plot.
}
\examples{
if (interactive()) {
  # Load in dummy results
  tpr <- TreatmentPatterns::TreatmentPatternsResults$new(
    filePath = system.file(
      "DummyOutput", "output.zip",
       package = "TreatmentPatterns"
     )
   )

  ggSunburst(
    treatmentPathways = tpr$treatment_pathways,
    minFreq = 20,
    style = "darwin"
  )
}
}
