% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOutput.R
\name{groupInfrequentCombinations}
\alias{groupInfrequentCombinations}
\title{groupInfrequentCombinations}
\usage{
groupInfrequentCombinations(data, groupCombinations)
}
\arguments{
\item{data}{Data}

\item{groupCombinations}{Group combinations}
}
\value{
data.table
}
\description{
Help function to group combinations
}
\examples{
\dontrun{
source(system.file(
  package = "TreatmentPatterns",
  "testing", "testParamsOutput.R"))
  
TreatmentPatterns:::groupInfrequentCombinations(
  data = treatment_pathways[[1]],
  groupCombinations = groupCombinations)
}
}
