% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSankeyDiagram.R
\name{groupInfrequentCombinations}
\alias{groupInfrequentCombinations}
\title{groupInfrequentCombinations}
\usage{
groupInfrequentCombinations(data, groupCombinations)
}
\arguments{
\item{data}{Data}

\item{groupCombinations}{Group combinations}
}
\value{
data.table
}
\description{
\code{DEPRECATED}\cr
Help function to group combinations
}
