% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSankeyDiagram.R
\name{groupInfrequentCombinations}
\alias{groupInfrequentCombinations}
\title{groupInfrequentCombinations}
\usage{
groupInfrequentCombinations(data, groupCombinations)
}
\arguments{
\item{data}{Data}

\item{groupCombinations}{(\code{logical(1)}: \code{FALSE})\cr
\describe{
\item{\code{TRUE}}{Group all combination treatments in category \code{"Combination"}.}
\item{\code{FALSE}}{Do not group combination treatments.}
}}
}
\value{
data.table
}
\description{
\code{DEPRECATED}\cr
Help function to group combinations
}
