% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCohort.R
\name{importCohorts}
\alias{importCohorts}
\title{This function will import the target and event cohorts following the path included in
this package if OMOP-CDM = FALSE.}
\usage{
importCohorts(cohortLocation, outputFolder)
}
\arguments{
\item{cohortLocation}{Location where cohorts are saved if OMOP-CDM = FALSE.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes (/).}
}
\description{
This function will import the target and event cohorts following the path included in
this package if OMOP-CDM = FALSE.
}
