% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{launchShinyApplication}
\alias{launchShinyApplication}
\title{Processes the zipped result folder(s) and launch the shiny application.}
\usage{
launchShinyApplication(zipFolder)
}
\arguments{
\item{zipFolder}{Name of local folder containing zip file(s).}
}
\value{

}
\description{
Processes the zipped result folder(s) and launch the shiny application.
}
