% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-PathwayConstructor.R
\name{PathwayConstructor}
\alias{PathwayConstructor}
\title{PathwayConstructor}
\description{
PathwayConstructor R6 object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PathwayConstructor-new}{\code{PathwayConstructor$new()}}
\item \href{#method-PathwayConstructor-validate}{\code{PathwayConstructor$validate()}}
\item \href{#method-PathwayConstructor-construct}{\code{PathwayConstructor$construct()}}
\item \href{#method-PathwayConstructor-getAndromeda}{\code{PathwayConstructor$getAndromeda()}}
\item \href{#method-PathwayConstructor-editSettings}{\code{PathwayConstructor$editSettings()}}
\item \href{#method-PathwayConstructor-getSettings}{\code{PathwayConstructor$getSettings()}}
\item \href{#method-PathwayConstructor-clone}{\code{PathwayConstructor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathwayConstructor-new"></a>}}
\if{latex}{\out{\hypertarget{method-PathwayConstructor-new}{}}}
\subsection{Method \code{new()}}{
Initialize method called by \code{PathwayConstructor$new()}.
\cr\cr
Choose the way you interface with the CDM, either through \code{DatabaseConnector} or \code{CDMConnector}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwayConstructor$new(cohorts, cohortTableName, cdmInterface)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cohorts}}{(\code{data.frame()})\cr
Cohort table.}

\item{\code{cohortTableName}}{(\code{character(1)})\cr
Cohort table name.}

\item{\code{cdmInterface}}{(\code{TreatmentPatterns::cdmInterface})\cr
A cdmInterface object created internally}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{invisible(self)})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathwayConstructor-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PathwayConstructor-validate}{}}}
\subsection{Method \code{validate()}}{
Validation method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwayConstructor$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{invisible(self)})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathwayConstructor-construct"></a>}}
\if{latex}{\out{\hypertarget{method-PathwayConstructor-construct}{}}}
\subsection{Method \code{construct()}}{
Construct the pathways. Generates \code{Andromeda::andromeda()} objects,
which can be fetched using \code{self$getAndromeda()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwayConstructor$construct()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathwayConstructor-getAndromeda"></a>}}
\if{latex}{\out{\hypertarget{method-PathwayConstructor-getAndromeda}{}}}
\subsection{Method \code{getAndromeda()}}{
Gets the \code{Andromeda::andromeda()} objects in a list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwayConstructor$getAndromeda()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{list()})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathwayConstructor-editSettings"></a>}}
\if{latex}{\out{\hypertarget{method-PathwayConstructor-editSettings}{}}}
\subsection{Method \code{editSettings()}}{
Edit settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwayConstructor$editSettings(
  studyName = "default",
  includeTreatments = "startDate",
  periodPriorToIndex = 0,
  minEraDuration = 0,
  splitEventCohorts = "",
  splitTime = 30,
  eraCollapseSize = 30,
  combinationWindow = 30,
  minPostCombinationDuration = 30,
  filterTreatments = "First",
  maxPathLength = 5,
  minCellCount = 5,
  minCellMethod = "Remove",
  groupCombinations = 10,
  addNoPaths = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{studyName}}{(\code{character(1)}: \code{"default"})\cr}

\item{\code{includeTreatments}}{(\code{character(1)}: \code{"startDate"})\cr}

\item{\code{periodPriorToIndex}}{(\code{integer(1)}: \code{0})\cr}

\item{\code{minEraDuration}}{(\code{integer(1)}: \code{0})\cr}

\item{\code{splitEventCohorts}}{(\code{character(n)}: \code{""})\cr}

\item{\code{splitTime}}{(\code{integer(1)}: \code{30})\cr}

\item{\code{eraCollapseSize}}{(\code{integer(1)}: \code{30})\cr}

\item{\code{combinationWindow}}{(\code{integer(1)}: \code{30})\cr}

\item{\code{minPostCombinationDuration}}{(\code{integer(1)}: \code{30})\cr}

\item{\code{filterTreatments}}{(\code{character(1)}: \code{"First"})\cr}

\item{\code{maxPathLength}}{(\code{integer(1)}: \code{5})\cr}

\item{\code{minCellCount}}{(\code{integer(1)}: \code{5})\cr}

\item{\code{minCellMethod}}{(\code{character(1)}: \code{"Remove"})\cr}

\item{\code{groupCombinations}}{(\code{integer(1)}: \code{10})\cr}

\item{\code{addNoPaths}}{(\code{logical(1)}: \code{TRUE})\cr}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{data.frame()})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathwayConstructor-getSettings"></a>}}
\if{latex}{\out{\hypertarget{method-PathwayConstructor-getSettings}{}}}
\subsection{Method \code{getSettings()}}{
Getter method to get the specified settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwayConstructor$getSettings()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{data.frame()})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathwayConstructor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PathwayConstructor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwayConstructor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
