% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-PathwayConstructor.R
\name{PathwayConstructor}
\alias{PathwayConstructor}
\title{PathwayConstructor}
\description{
PathwayConstructor

PathwayConstructor
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PathwayConstructor-new}{\code{PathwayConstructor$new()}}
\item \href{#method-PathwayConstructor-validate}{\code{PathwayConstructor$validate()}}
\item \href{#method-PathwayConstructor-construct}{\code{PathwayConstructor$construct()}}
\item \href{#method-PathwayConstructor-getAndromeda}{\code{PathwayConstructor$getAndromeda()}}
\item \href{#method-PathwayConstructor-clone}{\code{PathwayConstructor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathwayConstructor-new"></a>}}
\if{latex}{\out{\hypertarget{method-PathwayConstructor-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwayConstructor$new(
  cohortSettings,
  pathwaySettings,
  cohortTable,
  connectionDetails = NULL,
  cdmSchema = "",
  resultSchema = "",
  cdm = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathwayConstructor-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PathwayConstructor-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwayConstructor$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathwayConstructor-construct"></a>}}
\if{latex}{\out{\hypertarget{method-PathwayConstructor-construct}{}}}
\subsection{Method \code{construct()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwayConstructor$construct()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathwayConstructor-getAndromeda"></a>}}
\if{latex}{\out{\hypertarget{method-PathwayConstructor-getAndromeda}{}}}
\subsection{Method \code{getAndromeda()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwayConstructor$getAndromeda()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathwayConstructor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PathwayConstructor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwayConstructor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
