% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-PathwaySettings.R
\name{PathwaySettings}
\alias{PathwaySettings}
\title{PathwaySettings}
\description{
PathwaySettings

PathwaySettings
}
\section{Super class}{
\code{\link[TreatmentPatterns:Settings]{TreatmentPatterns::Settings}} -> \code{PathwaySettings}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PathwaySettings-addAnalysis}{\code{PathwaySettings$addAnalysis()}}
\item \href{#method-PathwaySettings-exportAnalyses}{\code{PathwaySettings$exportAnalyses()}}
\item \href{#method-PathwaySettings-importAnalyses}{\code{PathwaySettings$importAnalyses()}}
\item \href{#method-PathwaySettings-get}{\code{PathwaySettings$get()}}
\item \href{#method-PathwaySettings-validate}{\code{PathwaySettings$validate()}}
\item \href{#method-PathwaySettings-clone}{\code{PathwaySettings$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="TreatmentPatterns" data-topic="Settings" data-id="initialize"><a href='../../TreatmentPatterns/html/Settings.html#method-Settings-initialize'><code>TreatmentPatterns::Settings$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TreatmentPatterns" data-topic="Settings" data-id="print"><a href='../../TreatmentPatterns/html/Settings.html#method-Settings-print'><code>TreatmentPatterns::Settings$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathwaySettings-addAnalysis"></a>}}
\if{latex}{\out{\hypertarget{method-PathwaySettings-addAnalysis}{}}}
\subsection{Method \code{addAnalysis()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwaySettings$addAnalysis(analysisSettings)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathwaySettings-exportAnalyses"></a>}}
\if{latex}{\out{\hypertarget{method-PathwaySettings-exportAnalyses}{}}}
\subsection{Method \code{exportAnalyses()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwaySettings$exportAnalyses(filePath)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathwaySettings-importAnalyses"></a>}}
\if{latex}{\out{\hypertarget{method-PathwaySettings-importAnalyses}{}}}
\subsection{Method \code{importAnalyses()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwaySettings$importAnalyses(filePath, validate = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathwaySettings-get"></a>}}
\if{latex}{\out{\hypertarget{method-PathwaySettings-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwaySettings$get()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathwaySettings-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PathwaySettings-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwaySettings$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathwaySettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PathwaySettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwaySettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
