% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveAsFile.R
\name{saveAsFile}
\alias{saveAsFile}
\title{saveAsFile}
\usage{
saveAsFile(
  fileName,
  fileNameOut,
  zoom = 3,
  vwidth = 430,
  selector = "svg",
  ...
)
}
\arguments{
\item{fileName}{HTML-Filename.}

\item{fileNameOut}{Filename of image, either with .pdf or .png extension.}

\item{zoom}{Zoom factor, default = 3.}

\item{vwidth}{Width of frame to capture, default = 430.}

\item{selector}{Default: `"svg"`; svg only works if the output file is pdf, otherwise use `NULL`}

\item{...}{Other potential parameters to be passed to webshot2::webshot()}
}
\value{
Invisibly returns the normalized path to all screenshots taken. The character vector will have a class of '"webshot"'.
}
\description{
Save an SVG-image from a html-file as a PNG, - or PDF-file.
}
\examples{
\dontrun{
  saveAsFile("webFile.html", "outputFile.png")}
}
