% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructPathways.R
\name{selectRowsCombinationWindow}
\alias{selectRowsCombinationWindow}
\title{selectRowsCombinationWindow}
\usage{
selectRowsCombinationWindow(andromeda)
}
\arguments{
\item{andromeda}{(\code{Andromeda::andromeda()})}
}
\value{
(\code{invisible(NULL)})
}
\description{
Help function for doCombinationWindow that selects one overlapping drug era
per person to modify in next iteration of the combination window.
}
