% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{stratisfy}
\alias{stratisfy}
\title{stratisfy}
\usage{
stratisfy(treatmentHistory, years, ages, ageWindow, minFreq)
}
\arguments{
\item{treatmentHistory}{(\code{data.frame()})}

\item{years}{(\code{vector("character")})}

\item{ages}{(\code{vector("character")})}

\item{ageWindow}{(\code{integer(1)})}

\item{minFreq}{(\code{integer(1)})}
}
\value{
(\code{data.frame()})
}
\description{
stratisfy
}
