% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSunburstPlot.R
\name{stripname}
\alias{stripname}
\title{stripname}
\usage{
stripname(x, name)
}
\arguments{
\item{x}{input list}

\item{name}{the name of the list item from which the names will be removed}
}
\value{
list with removed names
}
\description{
Recursive function to remove name from all levels of list.
}
