% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOutput.R
\name{transformCSVtoJSON}
\alias{transformCSVtoJSON}
\title{transformCSVtoJSON}
\usage{
transformCSVtoJSON(data, outcomes, folder, file_name)
}
\arguments{
\item{data}{Data}

\item{outcomes}{Outcomes}

\item{folder}{Folder}

\item{file_name}{File name}
}
\value{
JSON
}
\description{
Help function to transform data in csv format to required JSON format for
HTML.
}
