% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSunburstPlot.R
\name{transformCSVtoJSON}
\alias{transformCSVtoJSON}
\title{transformCSVtoJSON}
\usage{
transformCSVtoJSON(data, outcomes)
}
\arguments{
\item{data}{(\code{data.frame()})}

\item{outcomes}{(\code{c()})}
}
\value{
the transformed csv as a json string
}
\description{
Help function to transform data in csv format to required JSON format for
HTML.
}
