% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatmentPatterns-package.R
\docType{package}
\name{TreatmentPatterns-package}
\alias{TreatmentPatterns}
\alias{TreatmentPatterns-package}
\title{TreatmentPatterns: A package to analyse real-world treatment patterns of a study population of interest}
\description{
This package gives insight in treatment patterns, the following tasks should be performed sequentially: 1. Construct treatment pathways. 2. Generate output (sunburst plots, Sankey diagrams and more). 3. Launch shiny application to visualize the results.
}
\keyword{internal}
