% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatmentPatterns-package.R
\docType{package}
\name{TreatmentPatterns-package}
\alias{TreatmentPatterns}
\alias{TreatmentPatterns-package}
\title{TreatmentPatterns: Analyzes Real-World Treatment Patterns of a Study Population of Interest}
\description{
Computes treatment patterns within a given cohort using the Observational Medical Outcomes Partnership (OMOP) common data model (CDM). As described in Markus, Verhamme, Kors, and Rijnbeek (2022) \doi{10.1016/j.cmpb.2022.107081}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/darwin-eu-dev/TreatmentPatterns}
  \item Report bugs at \url{https://github.com/darwin-eu-dev/TreatmentPatterns/issues}
}

}
\author{
\strong{Maintainer}: Maarten van Kessel \email{m.l.vankessel@erasmusmc.nl} (\href{https://orcid.org/0009-0006-8832-6030}{ORCID})

Authors:
\itemize{
  \item Aniek F Markus (\href{https://orcid.org/0000-0001-5779-4794}{ORCID})
}

}
\keyword{internal}
